[not verified] Removing Custom CSS Migration
