+++ projects/plugins/jetpack/modules/custom-css.php
@@ -16,24 +16,8 @@
  * Load custom CSS
  */
 function jetpack_load_custom_css() {
-	// If WordPress has the core version of Custom CSS, load our new version.
-	// @see https://core.trac.wordpress.org/changeset/38829
-	// @todo When should we remove the migration?
-	if ( function_exists( 'wp_get_custom_css' ) ) {
-		if ( ! function_exists( 'wp_update_custom_css_post' ) ) {
-			wp_die( 'Please run a SVN up to get the latest version of trunk, or update to at least 4.7 RC1' );
-		}
-		if ( ! Jetpack_Options::get_option( 'custom_css_4.7_migration' ) ) {
-			include_once __DIR__ . '/custom-css/migrate-to-core.php';
-		}
-
 		include_once __DIR__ . '/custom-css/custom-css/preprocessors.php';
-		include_once __DIR__ . '/custom-css/custom-css-4.7.php';
-		return;
-	}
-
-	include_once __DIR__ . '/custom-css/custom-css.php';
-	add_action( 'init', array( 'Jetpack_Custom_CSS', 'init' ) );
+		include_once __DIR__ . '/custom-css/custom-css.php';
 }
 
 add_action( 'jetpack_modules_loaded', 'custom_css_loaded' );
