+++ projects/plugins/jetpack/modules/custom-css/custom-css.php
@@ -1,926 +1,601 @@
 <?php // phpcs:ignore WordPress.Files.FileName.InvalidClassFileName
 
 use Automattic\Jetpack\Assets;
-use Automattic\Jetpack\Device_Detection\User_Agent_Info;
-use Automattic\Jetpack\Redirect;
 
 /**
- * Custom CSS class.
+ * Alternate Custom CSS source for 4.7 compat.
+ *
+ * @since 4.4.2
+ *
+ * @package automattic/jetpack
  */
-class Jetpack_Custom_CSS {
+
+/**
+ * Class Jetpack_Custom_CSS_Enhancements
+ */
+class Jetpack_Custom_CSS_Enhancements {
 
 	/**
-	 * Initialize the class.
+	 * Set up the actions and filters needed for our compatability layer on top of core's Custom CSS implementation.
 	 */
-	public static function init() {
-		add_action( 'switch_theme', array( __CLASS__, 'reset' ) );
-		add_action( 'wp_restore_post_revision', array( __CLASS__, 'restore_revision' ), 10, 2 );
+	public static function add_hooks() {
+		add_action( 'init', array( __CLASS__, 'init' ) );
+		add_action( 'customize_controls_enqueue_scripts', array( __CLASS__, 'customize_controls_enqueue_scripts' ) );
+		add_action( 'customize_register', array( __CLASS__, 'customize_register' ) );
+		add_filter( 'map_meta_cap', array( __CLASS__, 'map_meta_cap' ), 20, 2 );
+		add_action( 'customize_preview_init', array( __CLASS__, 'customize_preview_init' ) );
+		add_filter( '_wp_post_revision_fields', array( __CLASS__, 'wp_post_revision_fields' ), 10, 2 );
+		add_action( 'load-revision.php', array( __CLASS__, 'load_revision_php' ) );
 
-		// Save revisions for posts of type safecss.
-		add_action( 'load-revision.php', array( __CLASS__, 'add_revision_redirect' ) );
+		add_action( 'wp_enqueue_scripts', array( __CLASS__, 'wp_enqueue_scripts' ) );
 
-		// Override the edit link, the default link causes a redirect loop
-		add_filter( 'get_edit_post_link', array( __CLASS__, 'revision_post_link' ), 10, 3 );
+		// Handle Sass/LESS.
+		add_filter( 'customize_value_custom_css', array( __CLASS__, 'customize_value_custom_css' ), 10, 2 );
+		add_filter( 'customize_update_custom_css_post_content_args', array( __CLASS__, 'customize_update_custom_css_post_content_args' ), 10, 3 );
+		add_filter( 'update_custom_css_data', array( __CLASS__, 'update_custom_css_data' ), 10, 2 );
 
-		// Overwrite the content width global variable if one is set in the custom css
+		// Stuff for stripping out the theme's default stylesheet...
+		add_filter( 'stylesheet_uri', array( __CLASS__, 'style_filter' ) );
+		add_filter( 'safecss_skip_stylesheet', array( __CLASS__, 'preview_skip_stylesheet' ) );
+
+		// Stuff for overriding content width...
+		add_action( 'customize_preview_init', array( __CLASS__, 'preview_content_width' ) );
+		add_filter( 'jetpack_content_width', array( __CLASS__, 'jetpack_content_width' ) );
+		add_filter( 'editor_max_image_size', array( __CLASS__, 'editor_max_image_size' ), 10, 3 );
 		add_action( 'template_redirect', array( __CLASS__, 'set_content_width' ) );
 		add_action( 'admin_init', array( __CLASS__, 'set_content_width' ) );
 
-		if ( ! is_admin() ) {
-			add_filter( 'stylesheet_uri', array( __CLASS__, 'style_filter' ) );
-		}
+	/**
+	 * Things that we do on init.
+	 */
+	public static function init() {
 
-		define(
-			'SAFECSS_USE_ACE',
-			! jetpack_is_mobile() &&
-			! User_Agent_Info::is_ipad() &&
-			/**
-			 * Should the Custom CSS module use ACE to process CSS.
-			 *
-			 * @see https://ace.c9.io/
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 *
-			 * @param bool true Use ACE to process the Custom CSS. Default to true.
-			 */
-			apply_filters( 'safecss_use_ace', true )
+		wp_register_style( 'jetpack-codemirror', plugins_url( 'custom-css/css/codemirror.css', __FILE__ ), array(), '20120905' );
+		wp_register_style( 'jetpack-customizer-css', plugins_url( 'custom-css/css/customizer-control.css', __FILE__ ), array(), '20140728' );
+		wp_register_script( 'jetpack-codemirror', plugins_url( 'custom-css/js/codemirror.min.js', __FILE__ ), array(), '3.16', true );
+
+		$src = Assets::get_file_url_for_environment(
+			'_inc/build/custom-css/custom-css/js/core-customizer-css.core-4.9.min.js',
+			'modules/custom-css/custom-css/js/core-customizer-css.core-4.9.js'
 		);
-
-		// Register safecss as a custom post_type
-		// Explicit capability definitions are largely unnecessary because the posts are manipulated in code via an options page, managing CSS revisions does check the capabilities, so let's ensure that the proper caps are checked.
-		register_post_type(
-			'safecss',
+		wp_register_script(
+			'jetpack-customizer-css',
+			$src,
 			array(
-				/**
-				 * These are the defaults
-				 * 'exclude_from_search' => true,
-				 * 'public' => false,
-				 * 'publicly_queryable' => false,
-				 * 'show_ui' => false,
-				 */
-				'supports'     => array( 'revisions' ),
-				'label'        => 'Custom CSS',
-				'can_export'   => false,
-				'rewrite'      => false,
-				'capabilities' => array(
-					'edit_post'          => 'edit_theme_options',
-					'read_post'          => 'read',
-					'delete_post'        => 'edit_theme_options',
-					'edit_posts'         => 'edit_theme_options',
-					'edit_others_posts'  => 'edit_theme_options',
-					'publish_posts'      => 'edit_theme_options',
-					'read_private_posts' => 'read',
-				),
-			)
+				'customize-controls',
+				'underscore',
+			),
+			JETPACK__VERSION,
+			true
 		);
 
-		// Short-circuit WP if this is a CSS stylesheet request
-		if ( isset( $_GET['custom-css'] ) ) {
-			header( 'Content-Type: text/css', true, 200 );
-			header( 'Expires: ' . gmdate( 'D, d M Y H:i:s', time() + 31536000 ) . ' GMT' ); // 1 year
-			self::print_css();
-			exit;
-		}
+		wp_register_script(
+			'jetpack-customizer-css-preview',
+			Assets::get_file_url_for_environment(
+				'_inc/build/custom-css/custom-css/js/core-customizer-css-preview.min.js',
+				'modules/custom-css/custom-css/js/core-customizer-css-preview.js'
+			),
+			array( 'customize-selective-refresh' ),
+			JETPACK__VERSION,
+			true
+		);
 
-		add_action( 'admin_enqueue_scripts', array( 'Jetpack_Custom_CSS', 'enqueue_scripts' ) );
+		remove_action( 'wp_head', 'wp_custom_css_cb', 11 ); // 4.7.0 had it at 11, 4.7.1 moved it to 101.
+		remove_action( 'wp_head', 'wp_custom_css_cb', 101 );
+		add_action( 'wp_head', array( __CLASS__, 'wp_custom_css_cb' ), 101 );
 
-		if ( isset( $_GET['page'] ) && 'editcss' === $_GET['page'] && is_admin() ) {
-			// Do migration routine if necessary
-			self::upgrade();
-
-			/**
-			 * Allows additional work when migrating safecss from wp_options to wp_post.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 */
-			do_action( 'safecss_migrate_post' );
-		}
-
-		/**
-		 * Never embed the style in the head on wpcom.
-		 * Yes, this filter should be added to an unsynced file on wpcom, but
-		 * there is no good syntactically-correct location to put it yet.
-		 *
-		 * @link https://github.com/Automattic/jetpack/commit/a1be114e9179f64d147124727a58e2cf76c7e5a1#commitcomment-7763921
-		 */
-		if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
-			add_filter( 'safecss_embed_style', '__return_false' );
-		} else {
-			add_filter( 'safecss_embed_style', array( 'Jetpack_Custom_CSS', 'should_we_inline_custom_css' ), 10, 2 );
-		}
-
-		add_action( 'wp_head', array( 'Jetpack_Custom_CSS', 'link_tag' ), 101 );
-
-		add_filter( 'jetpack_content_width', array( 'Jetpack_Custom_CSS', 'jetpack_content_width' ) );
-		add_filter( 'editor_max_image_size', array( 'Jetpack_Custom_CSS', 'editor_max_image_size' ), 10, 3 );
-
-		if ( ! current_user_can( 'switch_themes' ) && ! is_super_admin() ) {
-			return;
-		}
-
-		add_action( 'admin_menu', array( 'Jetpack_Custom_CSS', 'menu' ) );
-
-		if ( isset( $_POST['safecss'] ) && ( ! isset( $_SERVER['REQUEST_URI'] ) || false === strstr( filter_var( wp_unslash( $_SERVER['REQUEST_URI'] ) ), 'options.php' ) ) ) {
-			check_admin_referer( 'safecss' );
-
-			$save_result = self::save(
-				array(
-					'css'             => filter_var( wp_unslash( $_POST['safecss'] ) ),
-					'is_preview'      => isset( $_POST['action'] ) && $_POST['action'] === 'preview',
-					'preprocessor'    => isset( $_POST['custom_css_preprocessor'] ) ? sanitize_key( $_POST['custom_css_preprocessor'] ) : '',
-					'add_to_existing' => isset( $_POST['add_to_existing'] ) ? $_POST['add_to_existing'] === 'true' : true,
-					'content_width'   => isset( $_POST['custom_content_width'] ) ? intval( $_POST['custom_content_width'] ) : false,
-				)
-			);
-
-			if ( $_POST['action'] === 'preview' ) {
-				wp_safe_redirect( add_query_arg( 'csspreview', 'true', get_option( 'home' ) ) );
-				exit;
-			}
-
-			if ( $save_result ) {
-				add_action( 'admin_notices', array( 'Jetpack_Custom_CSS', 'saved_message' ) );
-			}
-		}
-
-		// Modify all internal links so that preview state persists
-		if ( self::is_preview() ) {
-			ob_start( array( 'Jetpack_Custom_CSS', 'buffer' ) );
+		if ( isset( $_GET['custom-css'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- no changes made to the site.
+			self::print_linked_custom_css();
 		}
 	}
 
 	/**
-	 * Save new custom CSS. This should be the entry point for any third-party code using Jetpack_Custom_CSS
-	 * to save CSS.
-	 *
-	 * @param array $args Array of arguments:
-	 *        string $css The CSS (or LESS or Sass)
-	 *        bool $is_preview Whether this CSS is preview or published
-	 *        string preprocessor Which CSS preprocessor to use
-	 *        bool $add_to_existing Whether this CSS replaces the theme's CSS or supplements it.
-	 *        int $content_width A custom $content_width to go along with this CSS.
-	 * @return int The post ID of the saved Custom CSS post.
+	 * Things that we do on init when the Customize Preview is loading.
 	 */
-	public static function save( $args = array() ) {
-		$defaults = array(
-			'css'             => '',
-			'is_preview'      => false,
-			'preprocessor'    => '',
-			'add_to_existing' => true,
-			'content_width'   => false,
-		);
+	public static function customize_preview_init() {
+		add_filter( 'wp_get_custom_css', array( __CLASS__, 'customize_preview_wp_get_custom_css' ) );
+	}
 
-		$args = wp_parse_args( $args, $defaults );
+	/**
+	 * Print the current Custom CSS. This is for linking instead of printing directly.
+	 */
+	public static function print_linked_custom_css() {
+		header( 'Content-type: text/css' );
+		header( 'Expires: ' . gmdate( 'D, d M Y H:i:s', time() + YEAR_IN_SECONDS ) . ' GMT' );
+		echo wp_get_custom_css(); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		exit;
+	}
 
-		if (
-			$args['content_width']
-			&& (int) $args['content_width'] > 0
-			&& (
-				! isset( $GLOBALS['content_width'] )
-				|| $args['content_width'] !== $GLOBALS['content_width']
-			)
-		) {
-			$args['content_width'] = (int) $args['content_width'];
-		} else {
-			$args['content_width'] = false;
+	/**
+	 * Re-map the Edit CSS capability.
+	 *
+	 * Core, by default, restricts this to users that have `unfiltered_html` which
+	 * would make the feature unusable in multi-site by non-super-admins, due to Core
+	 * not shipping any solid sanitization.
+	 *
+	 * We're expanding who can use it, and then conditionally applying CSSTidy
+	 * sanitization to users that do not have the `unfiltered_html` capability.
+	 *
+	 * @param array  $caps Returns the user's actual capabilities.
+	 * @param string $cap  Capability name.
+	 *
+	 * @return array $caps
+	 */
+	public static function map_meta_cap( $caps, $cap ) {
+		if ( 'edit_css' === $cap ) {
+			$caps = array( 'edit_theme_options' );
+		}
+		return $caps;
+	}
+
+	/**
+	 * Shows Preprocessor code in the Revisions screen, and ensures that post_content_filtered
+	 * is maintained on revisions
+	 *
+	 * @param array $fields  Post fields pertinent to revisions.
+	 * @param array $post    A post array being processed for insertion as a post revision.
+	 *
+	 * @return array $fields Modified array to include post_content_filtered.
+	 */
+	public static function wp_post_revision_fields( $fields, $post ) {
+		// None of the fields in $post are required to be passed in this filter.
+		if ( ! isset( $post['post_type'], $post['ID'] ) ) {
+			return $fields;
 		}
 
-		/**
-		 * Fires prior to saving custom css values. Necessitated because the
-		 * core WordPress save_pre filters were removed:
-		 * - content_save_pre
-		 * - content_filtered_save_pre
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param array $args {
-		 * Array of custom CSS arguments.
-		 *  @type string $css The CSS (or LESS or Sass).
-		 *  @type bool $is_preview Whether this CSS is preview or published.
-		 *  @type string preprocessor Which CSS preprocessor to use.
-		 *  @type bool $add_to_existing Whether this CSS replaces the theme's CSS or supplements it.
-		 *  @type int $content_width A custom $content_width to go along with this CSS.
-		 * }
-		 */
-		do_action( 'safecss_save_pre', $args );
-
-		$warnings = array();
-
-		safecss_class();
-		$csstidy           = new csstidy();
-		$csstidy->optimise = new safecss( $csstidy );
-
-		$csstidy->set_cfg( 'remove_bslash', false );
-		$csstidy->set_cfg( 'compress_colors', false );
-		$csstidy->set_cfg( 'compress_font-weight', false );
-		$csstidy->set_cfg( 'optimise_shorthands', 0 );
-		$csstidy->set_cfg( 'remove_last_;', false );
-		$csstidy->set_cfg( 'case_properties', false );
-		$csstidy->set_cfg( 'discard_invalid_properties', true );
-		$csstidy->set_cfg( 'css_level', 'CSS3.0' );
-		$csstidy->set_cfg( 'preserve_css', true );
-		$csstidy->set_cfg( 'template', __DIR__ . '/csstidy/wordpress-standard.tpl' );
-
-		$prev = $args['css'];
-		$css  = preg_replace( '/\\\\([0-9a-fA-F]{4})/', '\\\\\\\\$1', $prev );
-		// prevent content: '\3434' from turning into '\\3434'
-		$css = str_replace( array( '\'\\\\', '"\\\\' ), array( '\'\\', '"\\' ), $css );
-
-		if ( $css !== $prev ) {
-			$warnings[] = 'preg_replace found stuff';
+		// If we're passed in a revision, go get the main post instead.
+		if ( 'revision' === $post['post_type'] ) {
+			$main_post_id = wp_is_post_revision( $post['ID'] );
+			$post         = get_post( $main_post_id, ARRAY_A );
 		}
-
-		// Some people put weird stuff in their CSS, KSES tends to be greedy
-		$css = str_replace( '<=', '&lt;=', $css );
-		// Why KSES instead of strip_tags?  Who knows?
-		$prev = $css;
-		$css  = wp_kses_split( $prev, array(), array() );
-		$css  = str_replace( '&gt;', '>', $css ); // kses replaces lone '>' with &gt;
-		// Why both KSES and strip_tags?  Because we just added some '>'.
-		$css = strip_tags( $css ); // phpcs:ignore WordPress.WP.AlternativeFunctions.strip_tags_strip_tags
-
-		if ( $css !== $prev ) {
-			$warnings[] = 'kses found stuff';
+		if ( 'custom_css' === $post['post_type'] ) {
+			$fields['post_content']          = __( 'CSS', 'jetpack' );
+			$fields['post_content_filtered'] = __( 'Preprocessor', 'jetpack' );
 		}
-
-		// if we're not using a preprocessor
-		if ( ! $args['preprocessor'] ) {
-
-			/**
-			 * Fires before parsing the css with CSSTidy, but only if
-			 * the preprocessor is not configured for use.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 *
-			 * @param obj $csstidy The csstidy object.
-			 * @param string $css Custom CSS.
-			 * @param array $args Array of custom CSS arguments.
-			 */
-			do_action( 'safecss_parse_pre', $csstidy, $css, $args );
-
-			$csstidy->parse( $css );
-
-			/**
-			 * Fires after parsing the css with CSSTidy, but only if
-			 * the preprocessor is not cinfigured for use.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 *
-			 * @param obj $csstidy The csstidy object.
-			 * @param array $warnings Array of warnings.
-			 * @param array $args Array of custom CSS arguments.
-			 */
-			do_action( 'safecss_parse_post', $csstidy, $warnings, $args );
-
-			$css = $csstidy->print->plain();
-		}
-
-		if ( $args['add_to_existing'] ) {
-			$add_to_existing = 'yes';
-		} else {
-			$add_to_existing = 'no';
-		}
-
-		if ( $args['is_preview'] || self::is_freetrial() ) {
-			// Save the CSS
-			$safecss_revision_id = self::save_revision( $css, true, $args['preprocessor'] );
-
-			// Cache Buster
-			update_option( 'safecss_preview_rev', (int) get_option( 'safecss_preview_rev' ) + 1 );
-
-			update_metadata( 'post', $safecss_revision_id, 'custom_css_add', $add_to_existing );
-			update_metadata( 'post', $safecss_revision_id, 'content_width', $args['content_width'] );
-			update_metadata( 'post', $safecss_revision_id, 'custom_css_preprocessor', $args['preprocessor'] );
-
-			delete_option( 'safecss_add' );
-			delete_option( 'safecss_content_width' );
-
-			if ( $args['is_preview'] ) {
-				return $safecss_revision_id;
-			}
-
-			/**
-			 * Fires after saving Custom CSS.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 */
-			do_action( 'safecss_save_preview_post' );
-		}
-
-		// Save the CSS
-		$safecss_post_id = self::save_revision( $css, false, $args['preprocessor'] );
-
-		$safecss_post_revision = self::get_current_revision();
-
-		update_option( 'safecss_rev', (int) get_option( 'safecss_rev' ) + 1 );
-
-		update_post_meta( $safecss_post_id, 'custom_css_add', $add_to_existing );
-		update_post_meta( $safecss_post_id, 'content_width', $args['content_width'] );
-		update_post_meta( $safecss_post_id, 'custom_css_preprocessor', $args['preprocessor'] );
-
-		delete_option( 'safecss_add' );
-		delete_option( 'safecss_content_width' );
-
-		update_metadata( 'post', $safecss_post_revision['ID'], 'custom_css_add', $add_to_existing );
-		update_metadata( 'post', $safecss_post_revision['ID'], 'content_width', $args['content_width'] );
-		update_metadata( 'post', $safecss_post_revision['ID'], 'custom_css_preprocessor', $args['preprocessor'] );
-
-		delete_option( 'safecss_preview_add' );
-
-		return $safecss_post_id;
+		return $fields;
 	}
 
 	/**
 	 * Get the published custom CSS post.
 	 *
-	 * @return array
+	 * @param string $stylesheet Optional. A theme object stylesheet name. Defaults to the current theme.
+	 * @return WP_Post|null
 	 */
-	public static function get_post() {
-		$custom_css_post_id = self::post_id();
-
-		if ( $custom_css_post_id ) {
-			return get_post( $custom_css_post_id, ARRAY_A );
-		}
-
-		return array();
+	public static function get_css_post( $stylesheet = '' ) {
+		return wp_get_custom_css_post( $stylesheet );
 	}
 
 	/**
-	 * Get the post ID of the published custom CSS post.
-	 *
-	 * @return int|bool The post ID if it exists; false otherwise.
+	 * Override Core's `wp_custom_css_cb` method to provide linking to custom css.
 	 */
-	public static function post_id() {
-		/**
-		 * Filter the ID of the post where Custom CSS is stored, before the ID is retrieved.
-		 *
-		 * If the callback function returns a non-null value, then post_id() will immediately
-		 * return that value, instead of retrieving the normal post ID.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 3.8.1
-		 *
-		 * @param null null The ID to return instead of the normal ID.
-		 */
-		$custom_css_post_id = apply_filters( 'jetpack_custom_css_pre_post_id', null );
-		if ( $custom_css_post_id !== null ) {
-			return $custom_css_post_id;
-		}
-
-		$custom_css_post_id = wp_cache_get( 'custom_css_post_id' );
-
-		if ( false === $custom_css_post_id ) {
-			$custom_css_posts = get_posts(
-				array(
-					'posts_per_page' => 1,
-					'post_type'      => 'safecss',
-					'post_status'    => 'publish',
-					'orderby'        => 'date',
-					'order'          => 'DESC',
-				)
-			);
-
-			if ( count( $custom_css_posts ) > 0 ) {
-				$custom_css_post_id = $custom_css_posts[0]->ID;
-			} else {
-				$custom_css_post_id = 0;
-			}
-
-			// Save post_id=0 to note that no safecss post exists.
-			wp_cache_set( 'custom_css_post_id', $custom_css_post_id );
-		}
-
-		if ( ! $custom_css_post_id ) {
-			return false;
-		}
-
-		return $custom_css_post_id;
-	}
-
-	/**
-	 * Get the current revision of the original safecss record
-	 *
-	 * @return object
-	 */
-	public static function get_current_revision() {
-		$safecss_post = self::get_post();
-
-		if ( empty( $safecss_post ) ) {
-			return false;
-		}
-
-		$revisions = wp_get_post_revisions(
-			$safecss_post['ID'],
-			array(
-				'posts_per_page' => 1,
-				'orderby'        => 'date',
-				'order'          => 'DESC',
-			)
-		);
-
-		// Empty array if no revisions exist
-		if ( empty( $revisions ) ) {
-			// Return original post
-			return $safecss_post;
-		} else {
-			// Return the first entry in $revisions, this will be the current revision
-			$current_revision = get_object_vars( array_shift( $revisions ) );
-			return $current_revision;
-		}
-	}
-
-	/**
-	 * Save new revision of CSS
-	 * Checks to see if content was modified before really saving
-	 *
-	 * @param string $css - the CSS.
-	 * @param bool   $is_preview - if we're in preview mode.
-	 * @param string $preprocessor - what preprocessor we're using.
-	 *
-	 * @return bool|int If nothing was saved, returns false. If a post
-	 *                  or revision was saved, returns the post ID.
-	 */
-	public static function save_revision( $css, $is_preview = false, $preprocessor = '' ) {
-		$safecss_post = self::get_post();
-
-		$compressed_css = self::minify( $css, $preprocessor );
-
-		// If null, there was no original safecss record, so create one
-		if ( ! $safecss_post ) {
-			if ( ! $css ) {
-				return false;
-			}
-
-			$post                          = array();
-			$post['post_content']          = wp_slash( $css );
-			$post['post_title']            = 'safecss';
-			$post['post_status']           = 'publish';
-			$post['post_type']             = 'safecss';
-			$post['post_content_filtered'] = wp_slash( $compressed_css );
-
-			// Set excerpt to current theme, for display in revisions list
-			$current_theme        = wp_get_theme();
-			$post['post_excerpt'] = $current_theme->Name; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
-
-			add_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9, 3 );
-			// Insert the CSS into wp_posts
-			$post_id = wp_insert_post( $post );
-			remove_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9 );
-			wp_cache_set( 'custom_css_post_id', $post_id );
-			return $post_id;
-		}
-
-		// Update CSS in post array with new value passed to this function
-		$safecss_post['post_content']          = $css;
-		$safecss_post['post_content_filtered'] = $compressed_css;
-
-		// Set excerpt to current theme, for display in revisions list
-		$current_theme                = wp_get_theme();
-		$safecss_post['post_excerpt'] = $current_theme->Name; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
-
-		// Don't carry over last revision's timestamps, otherwise revisions all have matching timestamps
-		unset( $safecss_post['post_date'] );
-		unset( $safecss_post['post_date_gmt'] );
-		unset( $safecss_post['post_modified'] );
-		unset( $safecss_post['post_modified_gmt'] );
-
-		// Do not update post if we are only saving a preview
-		if ( false === $is_preview ) {
-			$safecss_post['post_content']          = wp_slash( $safecss_post['post_content'] );
-			$safecss_post['post_content_filtered'] = wp_slash( $safecss_post['post_content_filtered'] );
-			add_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9, 3 );
-			$post_id = wp_update_post( $safecss_post );
-			remove_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9 );
-			wp_cache_set( 'custom_css_post_id', $post_id );
-			return $post_id;
-		} elseif ( ! defined( 'DOING_MIGRATE' ) ) {
-			add_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9, 3 );
-			$revision = _wp_put_post_revision( $safecss_post );
-			remove_filter( 'wp_insert_post_data', array( __CLASS__, 'restore_unsafe_postcss_content' ), 9 );
-			return $revision;
-		}
-	}
-
-	/**
-	 * Restore Unsafe Post CSS Content.
-	 *
-	 * @param array $data The post data being filtered.
-	 * @param array $postarray Unused.
-	 * @param array $unsanitized The unsanitized post data.
-	 *
-	 * @return array Post data.
-	 */
-	public static function restore_unsafe_postcss_content( $data, $postarray, $unsanitized ) {
-		$replace_content =
-				isset( $data['post_type'] ) &&
-				isset( $unsanitized['post_content'] ) &&
-				(
-						'safecss' === $data['post_type'] ||
-						(
-								'revision' === $data['post_type'] &&
-								! empty( $data['post_parent'] ) &&
-								'safecss' === get_post_type( $data['post_parent'] )
-						)
-				);
-		if ( $replace_content ) {
-			$data['post_content'] = $unsanitized['post_content'];
-		}
-		return $data;
-	}
-
-	/**
-	 * Prevent the stylesheet from being enqued.
-	 *
-	 * @return bool
-	 */
-	public static function skip_stylesheet() {
-		/**
-		 * Prevent the Custom CSS stylesheet from being enqueued.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 2.2.1
-		 *
-		 * @param null Should the stylesheet be skipped. Default to null. Anything else will force the stylesheet to be skipped.
-		 */
-		$skip_stylesheet = apply_filters( 'safecss_skip_stylesheet', null );
-
-		if ( null !== $skip_stylesheet ) {
-			return $skip_stylesheet;
-		} elseif ( self::is_customizer_preview() ) {
-			return false;
-		} else {
-			if ( self::is_preview() ) {
-				$safecss_post = self::get_current_revision();
-
-				if ( $safecss_post ) {
-					return (bool) ( get_post_meta( $safecss_post['ID'], 'custom_css_add', true ) === 'no' );
-				} else {
-					return (bool) ( get_option( 'safecss_preview_add' ) === 'no' );
-				}
-			} else {
-				$custom_css_post_id = self::post_id();
-
-				if ( $custom_css_post_id ) {
-					$custom_css_add = get_post_meta( $custom_css_post_id, 'custom_css_add', true );
-
-					// It is possible for the CSS to be stored in a post but for the safecss_add option
-					// to have not been upgraded yet if the user hasn't opened their Custom CSS editor
-					// since October 2012.
-					if ( ! empty( $custom_css_add ) ) {
-						return (bool) ( $custom_css_add === 'no' );
-					}
-				}
-
-				return (bool) ( Jetpack_Options::get_option_and_ensure_autoload( 'safecss_add', '' ) === 'no' );
-			}
-		}
-	}
-
-	/**
-	 * Checks if we're in a preview mode.
-	 *
-	 * @return bool
-	 */
-	public static function is_preview() {
-		return isset( $_GET['csspreview'] ) && $_GET['csspreview'] === 'true'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- no changes made to the site.
-	}
-
-	/**
-	 * Currently this filter function gets called on
-	 * 'template_redirect' action and
-	 * 'admin_init' action
-	 */
-	public static function set_content_width() {
-		// Don't apply this filter on the Edit CSS page
-		if ( isset( $_GET ) && isset( $_GET['page'] ) && 'editcss' === $_GET['page'] && is_admin() ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- nothing changing on the site, it's not applying a filter if set.
+	public static function wp_custom_css_cb() {
+		$styles = wp_get_custom_css();
+		if ( ! $styles ) {
 			return;
 		}
 
-		$GLOBALS['content_width'] = Jetpack::get_content_width();
+		$should_embed = strlen( $styles ) < 2000;
+		/** This filter is documented in projects/plugins/jetpack/modules/custom-css/custom-css.php */
+		$should_embed = apply_filters( 'safecss_embed_style', $should_embed, $styles );
+
+		if ( $should_embed || is_customize_preview() ) {
+			printf(
+				'<style type="text/css" id="wp-custom-css">%1$s</style>',
+				wp_strip_all_tags( $styles ) // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+			);
+		} else {
+			// Add a cache buster to the url.
+			$url = home_url( '/' );
+			$url = add_query_arg( 'custom-css', substr( md5( $styles ), -10 ), $url );
+
+			printf(
+				'<link rel="stylesheet" type="text/css" id="wp-custom-css" href="%1$s" />', // phpcs:ignore WordPress.WP.EnqueuedResources.NonEnqueuedStylesheet
+				esc_url( $url )
+			);
+		}
 	}
 
 	/**
-	 * False when the site has the Custom Design upgrade.
-	 * Used only on WordPress.com.
+	 * Get the ID of a Custom CSS post tying to a given stylesheet.
 	 *
-	 * @return bool
-	 * @todo see if we can remove this, I don't believe WordPress.com uses free trials anymore.
+	 * @param string $stylesheet Stylesheet name.
+	 *
+	 * @return int $post_id Post ID.
 	 */
-	public static function is_freetrial() {
-		/**
-		 * Determine if a WordPress.com site uses a Free trial of the Custom Design Upgrade.
-		 * Used only on WordPress.com.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param bool false Does the site use a Free trial of the Custom Design Upgrade. Default to false.
-		 */
-		return apply_filters( 'safecss_is_freetrial', false );
+	public static function post_id( $stylesheet = '' ) {
+		$post = self::get_css_post( $stylesheet );
+		if ( $post instanceof WP_Post ) {
+			return $post->ID;
+		}
+		return 0;
 	}
 
 	/**
-	 * Get the preprocessor key.
-	 *
-	 * @return string|false
+	 * Partial for use in the Customizer.
 	 */
-	public static function get_preprocessor_key() {
-		$safecss_post = self::get_current_revision();
-		return get_post_meta( $safecss_post['ID'], 'custom_css_preprocessor', true );
+	public static function echo_custom_css_partial() {
+		echo wp_get_custom_css(); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 	}
 
 	/**
-	 * Get the prepocessor.
+	 * Admin page!
 	 *
-	 * @return string|null
+	 * This currently has two main uses -- firstly to display the css for an inactive
+	 * theme if there are no revisions attached it to a legacy bug, and secondly to
+	 * handle folks that have bookmarkes in their browser going to the old page for
+	 * managing Custom CSS in Jetpack.
+	 *
+	 * If we ever add back in a non-Customizer CSS editor, this would be the place.
 	 */
-	public static function get_preprocessor() {
-		/** This filter is documented in modules/custom-css/custom-css.php */
-		$preprocessors             = apply_filters( 'jetpack_custom_css_preprocessors', array() );
-		$selected_preprocessor_key = self::get_preprocessor_key();
-		$selected_preprocessor     = isset( $preprocessors[ $selected_preprocessor_key ] ) ? $preprocessors[ $selected_preprocessor_key ] : null;
-		return $selected_preprocessor;
+	public static function admin_page() {
+		$post       = null;
+		$stylesheet = null;
+		if ( isset( $_GET['id'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- no changes made to the site.
+			$post_id = absint( $_GET['id'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- no changes made to the site
+			$post    = get_post( $post_id );
+			if ( $post instanceof WP_Post && 'custom_css' === $post->post_type ) {
+				$stylesheet = $post->post_title;
+			}
+		}
+		?>
+		<div class="wrap">
+			<?php self::revisions_switcher_box( $stylesheet ); ?>
+			<h1>
+				<?php
+				if ( $post ) {
+					printf( 'Custom CSS for &#8220;%1$s&#8221;', wp_get_theme( $stylesheet )->Name ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+				} else {
+					esc_html_e( 'Custom CSS', 'jetpack' );
+				}
+				if ( current_user_can( 'customize' ) ) {
+					printf(
+						' <a class="page-title-action hide-if-no-customize" href="%1$s">%2$s</a>',
+						esc_url( self::customizer_link() ),
+						esc_html__( 'Manage with Live Preview', 'jetpack' )
+					);
+				}
+				?>
+			</h1>
+			<p><?php esc_html_e( 'Custom CSS is now managed in the Customizer.', 'jetpack' ); ?></p>
+			<?php if ( $post ) : ?>
+				<div class="revisions">
+					<h3><?php esc_html_e( 'CSS', 'jetpack' ); ?></h3>
+					<textarea class="widefat" readonly><?php echo esc_textarea( $post->post_content ); ?></textarea>
+					<?php if ( $post->post_content_filtered ) : ?>
+						<h3><?php esc_html_e( 'Preprocessor', 'jetpack' ); ?></h3>
+						<textarea class="widefat" readonly><?php echo esc_textarea( $post->post_content_filtered ); ?></textarea>
+					<?php endif; ?>
+				</div>
+			<?php endif; ?>
+		</div>
+
+		<style>
+			.other-themes-wrap {
+				float: right;
+				background-color: #fff;
+				-webkit-box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+				box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+				padding: 5px 10px;
+				margin-bottom: 10px;
+			}
+			.other-themes-wrap label {
+				display: block;
+				margin-bottom: 10px;
+			}
+			.other-themes-wrap select {
+				float: left;
+				width: 77%;
+			}
+			.other-themes-wrap button {
+				float: right;
+				width: 20%;
+			}
+			.revisions {
+				clear: both;
+			}
+			.revisions textarea {
+				min-height: 300px;
+				background: #fff;
+			}
+		</style>
+		<script>
+			(function($){
+				var $switcher = $('.other-themes-wrap');
+				$switcher.find('button').on('click', function(e){
+					e.preventDefault();
+					if ( $switcher.find('select').val() ) {
+						window.location.href = $switcher.find('select').val();
+					}
+				});
+			})(jQuery);
+		</script>
+		<?php
 	}
 
 	/**
-	 * Get the CSS.
+	 * Build the URL to deep link to the Customizer.
 	 *
-	 * @param boolean $compressed - if the CSS is compressed.
+	 * You can modify the return url via $args.
 	 *
+	 * @param array $args Array of parameters.
 	 * @return string
 	 */
-	public static function get_css( $compressed = false ) {
-		/**
-		 * Filter the Custom CSS returned.
-		 * Can be used to return an error, or no CSS at all.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param bool false Should we return an error instead of the Custom CSS. Default to false.
-		 */
-		$default_css = apply_filters( 'safecss_get_css_error', false );
+	public static function customizer_link( $args = array() ) {
+		$args = wp_parse_args(
+			$args,
+			array(
+				'return_url' => isset( $_SERVER['REQUEST_URI'] ) ? rawurlencode( filter_var( wp_unslash( $_SERVER['REQUEST_URI'] ) ) ) : '',
+			)
+		);
 
-		if ( $default_css !== false ) {
-			return $default_css;
+		return add_query_arg(
+			array(
+				array(
+					'autofocus' => array(
+						'section' => 'custom_css',
+					),
+				),
+				'return' => $args['return_url'],
+			),
+			admin_url( 'customize.php' )
+		);
+	}
+
+	/**
+	 * Handle the enqueueing and localizing for scripts to be used in the Customizer.
+	 */
+	public static function customize_controls_enqueue_scripts() {
+		wp_enqueue_style( 'jetpack-customizer-css' );
+		wp_enqueue_script( 'jetpack-customizer-css' );
+
+		$content_help = __( 'Set a different media width for full size images.', 'jetpack' );
+		if ( ! empty( $GLOBALS['content_width'] ) ) {
+			$content_help .= sprintf(
+				// translators: the theme name and then the default width.
+				_n( ' The default media width for the <strong>%1$s</strong> theme is %2$d pixel.', ' The default media width for the <strong>%1$s</strong> theme is %2$d pixels.', (int) $GLOBALS['content_width'], 'jetpack' ),
+				wp_get_theme()->Name,
+				(int) $GLOBALS['content_width']
+			);
 		}
 
-		$option = ( self::is_preview() || self::is_freetrial() ) ? 'safecss_preview' : 'safecss';
-		$css    = '';
+		wp_localize_script(
+			'jetpack-customizer-css',
+			'_jp_css_settings',
+			array(
+				/** This filter is documented in modules/custom-css/custom-css.php */
+				'useRichEditor'        => ! jetpack_is_mobile() && apply_filters( 'safecss_use_ace', true ),
+				'areThereCssRevisions' => self::are_there_css_revisions(),
+				'revisionsUrl'         => self::get_revisions_url(),
+				'cssHelpUrl'           => '//en.support.wordpress.com/custom-design/editing-css/',
+				'l10n'                 => array(
+					'mode'           => __( 'Start Fresh', 'jetpack' ),
+					'mobile'         => __( 'On Mobile', 'jetpack' ),
+					'contentWidth'   => $content_help,
+					'revisions'      => _x( 'See full history', 'Toolbar button to see full CSS revision history', 'jetpack' ),
+					'css_help_title' => _x( 'Help', 'Toolbar button to get help with custom CSS', 'jetpack' ),
+				),
+			)
+		);
+	}
 
-		if ( 'safecss' === $option ) {
-			// Don't bother checking for a migrated 'safecss' option if it never existed.
-			if ( false === get_option( 'safecss' ) || get_option( 'safecss_revision_migrated' ) ) {
-				$safecss_post = self::get_post();
-				if ( ! empty( $safecss_post ) ) {
-					$css = ( $compressed && $safecss_post['post_content_filtered'] ) ? $safecss_post['post_content_filtered'] : $safecss_post['post_content'];
-				}
-			} else {
-				$current_revision = self::get_current_revision();
-				if ( false === $current_revision ) {
-					$css = '';
-				} else {
-					$css = ( $compressed && $current_revision['post_content_filtered'] ) ? $current_revision['post_content_filtered'] : $current_revision['post_content'];
-				}
-			}
+	/**
+	 * Check whether there are CSS Revisions for a given theme.
+	 *
+	 * Going forward, there should always be, but this was necessitated
+	 * early on by https://core.trac.wordpress.org/ticket/30854
+	 *
+	 * @param string $stylesheet Stylesheet name.
+	 *
+	 * @return bool|null|WP_Post
+	 */
+	public static function are_there_css_revisions( $stylesheet = '' ) {
+		$post = wp_get_custom_css_post( $stylesheet );
+		if ( empty( $post ) ) {
+			return $post;
+		}
+		return (bool) wp_get_post_revisions( $post );
+	}
 
-			// Fix for un-migrated Custom CSS
-			if ( empty( $safecss_post ) ) {
-				$_css = get_option( 'safecss' );
-				if ( ! empty( $_css ) ) {
-					$css = $_css;
-				}
+	/**
+	 * Core doesn't have a function to get the revisions url for a given post ID.
+	 *
+	 * @param string $stylesheet Stylesheet name.
+	 *
+	 * @return null|string|void
+	 */
+	public static function get_revisions_url( $stylesheet = '' ) {
+		$post = wp_get_custom_css_post( $stylesheet );
+
+		// If we have any currently saved customizations...
+		if ( $post instanceof WP_Post ) {
+			$revisions = wp_get_post_revisions( $post->ID, array( 'posts_per_page' => 1 ) );
+			if ( empty( $revisions ) || is_wp_error( $revisions ) ) {
+				return admin_url( 'themes.php?page=editcss' );
 			}
-		} elseif ( 'safecss_preview' === $option ) {
-			$safecss_post = self::get_current_revision();
-			$css          = $safecss_post['post_content'];
-			$css          = self::minify( $css, get_post_meta( $safecss_post['ID'], 'custom_css_preprocessor', true ) );
+			$revision = reset( $revisions );
+			return get_edit_post_link( $revision->ID );
 		}
 
-		$css = str_replace( array( '\\\00BB \\\0020', '\0BB \020', '0BB 020' ), '\00BB \0020', $css );
+		return admin_url( 'themes.php?page=editcss' );
+	}
 
-		/**
-		 * Filter the Custom CSS returned from the editor.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param string $css Custom CSS.
-		 */
-		$css = apply_filters( 'safecss_css', $css );
+	/**
+	 * Get a map of all theme names and theme stylesheets for mapping stuff.
+	 *
+	 * @return array
+	 */
+	public static function get_themes() {
+		$themes = wp_get_themes( array( 'errors' => null ) );
+		$all    = array();
+		foreach ( $themes as $theme ) {
+			$all[ $theme->name ] = $theme->stylesheet;
+		}
+		return $all;
+	}
 
+	/**
+	 * When we need to get all themes that have Custom CSS saved.
+	 *
+	 * @return array
+	 */
+	public static function get_all_themes_with_custom_css() {
+		$themes     = self::get_themes();
+		$custom_css = get_posts(
+			array(
+				'post_type'   => 'custom_css',
+				'post_status' => get_post_stati(),
+				'number'      => -1,
+				'order'       => 'DESC',
+				'orderby'     => 'modified',
+			)
+		);
+		$return     = array();
+
+		foreach ( $custom_css as $post ) {
+			$stylesheet = $post->post_title;
+			$label      = array_search( $stylesheet, $themes, true );
+
+			if ( ! $label ) {
+				continue;
+			}
+
+			$return[ $stylesheet ] = array(
+				'label' => $label,
+				'post'  => $post,
+			);
+		}
+
+		return $return;
+	}
+
+	/**
+	 * Handle the enqueueing of scripts for customize previews.
+	 */
+	public static function wp_enqueue_scripts() {
+		if ( is_customize_preview() ) {
+			wp_enqueue_script( 'jetpack-customizer-css-preview' );
+			wp_localize_script(
+				'jetpack-customizer-css-preview',
+				'jpCustomizerCssPreview',
+				array(
+					/** This filter is documented in modules/custom-css/custom-css.php */
+					'preprocessors' => apply_filters( 'jetpack_custom_css_preprocessors', array() ),
+				)
+			);
+		}
+	}
+
+	/**
+	 * Sanitize the CSS for users without `unfiltered_html`.
+	 *
+	 * @param string $css  Input CSS.
+	 * @param array  $args Array of CSS options.
+	 *
+	 * @return mixed|string
+	 */
+	public static function sanitize_css( $css, $args = array() ) {
+		$args = wp_parse_args(
+			$args,
+			array(
+				'force'        => false,
+				'preprocessor' => null,
+			)
+		);
+
+		if ( $args['force'] || ! current_user_can( 'unfiltered_html' ) ) {
+
+			$warnings = array();
+
+			safecss_class();
+			$csstidy           = new csstidy();
+			$csstidy->optimise = new safecss( $csstidy );
+
+			$csstidy->set_cfg( 'remove_bslash', false );
+			$csstidy->set_cfg( 'compress_colors', false );
+			$csstidy->set_cfg( 'compress_font-weight', false );
+			$csstidy->set_cfg( 'optimise_shorthands', 0 );
+			$csstidy->set_cfg( 'remove_last_;', false );
+			$csstidy->set_cfg( 'case_properties', false );
+			$csstidy->set_cfg( 'discard_invalid_properties', true );
+			$csstidy->set_cfg( 'css_level', 'CSS3.0' );
+			$csstidy->set_cfg( 'preserve_css', true );
+			$csstidy->set_cfg( 'template', __DIR__ . '/csstidy/wordpress-standard.tpl' );
+
+			// Test for some preg_replace stuff.
+			$prev = $css;
+			$css  = preg_replace( '/\\\\([0-9a-fA-F]{4})/', '\\\\\\\\$1', $css );
+			// prevent content: '\3434' from turning into '\\3434'.
+			$css = str_replace( array( '\'\\\\', '"\\\\' ), array( '\'\\', '"\\' ), $css );
+			if ( $css !== $prev ) {
+				$warnings[] = 'preg_replace found stuff';
+			}
+
+			// Some people put weird stuff in their CSS, KSES tends to be greedy.
+			$css = str_replace( '<=', '&lt;=', $css );
+
+			// Test for some kses stuff.
+			$prev = $css;
+			// Why KSES instead of strip_tags?  Who knows?
+			$css = wp_kses_split( $css, array(), array() );
+			$css = str_replace( '&gt;', '>', $css ); // kses replaces lone '>' with &gt;
+			// Why both KSES and strip_tags?  Because we just added some '>'.
+			$css = strip_tags( $css ); // phpcs:ignore WordPress.WP.AlternativeFunctions.strip_tags_strip_tags -- scared to update this to wp_strip_all_tags since we're building a CSS file here.
+
+			if ( $css !== $prev ) {
+				$warnings[] = 'kses found stuff';
+			}
+
+			// if we're not using a preprocessor.
+			if ( ! $args['preprocessor'] ) {
+
+				/** This action is documented in modules/custom-css/custom-css.php */
+				do_action( 'safecss_parse_pre', $csstidy, $css, $args );
+
+				$csstidy->parse( $css );
+
+				/** This action is documented in modules/custom-css/custom-css.php */
+				do_action( 'safecss_parse_post', $csstidy, $warnings, $args );
+
+				$css = $csstidy->print->plain();
+			}
+		}
 		return $css;
 	}
 
 	/**
-	 * Replace insecure URLs.
-	 *
-	 * @param string $css - the CSS.
-	 *
-	 * @return string
+	 * Override $content_width in customizer previews.
 	 */
-	public static function replace_insecure_urls( $css ) {
-		if ( ! function_exists( '_sa_get_frontend_https_url_replacement_map' ) ) {
-			return $css;
-		}
-		list( $http_urls, $secure_urls ) = _sa_get_frontend_https_url_replacement_map();
-
-		return str_replace( $http_urls, $secure_urls, $css );
-	}
-
-	/**
-	 * Print the CSS.
-	 */
-	public static function print_css() {
-
-		/**
-		 * Fires right before printing the custom CSS inside the <head> element.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 */
-		do_action( 'safecss_print_pre' );
-		$css = self::get_css( true );
-		echo self::replace_insecure_urls( $css ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-	}
-
-	/**
-	 * If the CSS is less than 2,000 characters, inline it! otherwise return what was passed in.
-	 *
-	 * @param bool   $should_we if we should inline the CSS.
-	 * @param string $css - the CSS object.
-	 *
-	 * @return bool
-	 */
-	public static function should_we_inline_custom_css( $should_we, $css ) {
-		return ( strlen( $css ) < 2000 ) ? true : $should_we;
-	}
-
-	/**
-	 * Add the link tag to inline CSS.
-	 */
-	public static function link_tag() {
-		global $blog_id, $current_blog;
-
-		if (
-			/**
-			 * Do not include any CSS on the page if the CSS includes an error.
-			 * Setting this filter to true stops any Custom CSS from being enqueued.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 *
-			 * @param bool false Does the CSS include an error. Default to false.
-			 */
-			apply_filters( 'safecss_style_error', false )
-		) {
+	public static function preview_content_width() {
+		global $wp_customize;
+		if ( ! is_customize_preview() ) {
 			return;
 		}
 
-		if ( ! is_super_admin() && isset( $current_blog ) && ( $current_blog->spam || $current_blog->deleted ) ) {
+		$setting = $wp_customize->get_setting( 'jetpack_custom_css[content_width]' );
+		if ( ! $setting ) {
 			return;
 		}
 
-		if ( self::is_customizer_preview() ) {
-			return;
+		$customized_content_width = (int) $setting->post_value();
+		if ( ! empty( $customized_content_width ) ) {
+			$GLOBALS['content_width'] = $customized_content_width;
 		}
-
-		$css    = '';
-		$option = self::is_preview() ? 'safecss_preview' : 'safecss';
-
-		if ( 'safecss' === $option ) {
-			if ( Jetpack_Options::get_option_and_ensure_autoload( 'safecss_revision_migrated', '0' ) ) {
-				$safecss_post = self::get_post();
-
-				if ( ! empty( $safecss_post['post_content'] ) ) {
-					$css = $safecss_post['post_content'];
-				}
-			} else {
-				$current_revision = self::get_current_revision();
-
-				if ( ! empty( $current_revision['post_content'] ) ) {
-					$css = $current_revision['post_content'];
-				}
-			}
-
-			// Fix for un-migrated Custom CSS
-			if ( empty( $safecss_post ) ) {
-				$_css = Jetpack_Options::get_option_and_ensure_autoload( 'safecss', '' );
-				if ( ! empty( $_css ) ) {
-					$css = $_css;
-				}
-			}
-		}
-
-		if ( 'safecss_preview' === $option ) {
-			$safecss_post = self::get_current_revision();
-
-			if ( ! empty( $safecss_post['post_content'] ) ) {
-				$css = $safecss_post['post_content'];
-			}
-		}
-
-		$css = str_replace( array( '\\\00BB \\\0020', '\0BB \020', '0BB 020' ), '\00BB \0020', $css );
-
-		if ( $css === '' ) {
-			return;
-		}
-
-		if (
-			/**
-			 * Allow inserting CSS inline instead of through a separate file.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 3.4.0
-			 *
-			 * @param bool false Should the CSS be added inline instead of through a separate file. Default to false.
-			 * @param string $css Custom CSS.
-			 */
-			apply_filters( 'safecss_embed_style', false, $css )
-		) {
-
-			echo "\r\n" . '<style id="custom-css-css">' . self::get_css( true ) . "</style>\r\n"; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-
-		} else {
-
-			$href = home_url( '/' );
-			$href = add_query_arg( 'custom-css', 1, $href );
-			$href = add_query_arg( 'csblog', $blog_id, $href );
-			$href = add_query_arg( 'cscache', 6, $href );
-			$href = add_query_arg( 'csrev', (int) get_option( $option . '_rev' ), $href );
-
-			/**
-			 * Filter the Custom CSS link enqueued in the head.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 *
-			 * @param string $href Custom CSS link enqueued in the head.
-			 * @param string $blog_id Blog ID.
-			 */
-			$href = apply_filters( 'safecss_href', $href, $blog_id );
-
-			if ( self::is_preview() ) {
-				$href = add_query_arg( 'csspreview', 'true', $href );
-			}
-
-			?>
-			<link rel="stylesheet" id="custom-css-css" type="text/css" href="<?php echo esc_url( $href ); // phpcs:ignore WordPress.WP.EnqueuedResources.NonEnqueuedStylesheet ?>" />
-			<?php
-
-		}
-
-		/**
-		 * Fires after creating the <link> in the <head> element for the custom css stylesheet.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 2.2.2
-		 */
-		do_action( 'safecss_link_tag_post' );
 	}
 
 	/**
-	 * Filter the default blank Custom CSS URL.
+	 * Filter the current theme's stylesheet for potentially nullifying it.
 	 *
-	 * @param string $current - the current CSS.
+	 * @param string $current Stylesheet URI for the current theme/child theme.
 	 *
-	 * @return string
+	 * @return mixed|void
 	 */
 	public static function style_filter( $current ) {
-		if ( self::is_freetrial() && ( ! self::is_preview() || ! current_user_can( 'switch_themes' ) ) ) {
+		if ( is_admin() ) {
+			return $current;
+		} elseif ( self::is_freetrial() && ( ! self::is_preview() || ! current_user_can( 'switch_themes' ) ) ) {
 			return $current;
 		} elseif ( self::skip_stylesheet() ) {
-			/**
-			 * Filter the default blank Custom CSS URL.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 2.2.1
-			 *
-			 * @param string $url Default blank Custom CSS URL.
-			 */
+			/** This filter is documented in modules/custom-css/custom-css.php */
 			return apply_filters( 'safecss_style_filter_url', plugins_url( 'custom-css/css/blank.css', __FILE__ ) );
 		}
 
@@ -928,868 +603,360 @@ class Jetpack_Custom_CSS {
 	}
 
 	/**
-	 * Buffer the HTML.
+	 * Determine whether or not we should have the theme skip its main stylesheet.
 	 *
-	 * @param string $html - the HTML.
-	 *
-	 * @return string
+	 * @return mixed The truthiness of this value determines whether the stylesheet should be skipped.
 	 */
-	public static function buffer( $html ) {
-		$html = str_replace( '</body>', self::preview_flag(), $html );
-		return preg_replace_callback( '!href=([\'"])(.*?)\\1!', array( 'Jetpack_Custom_CSS', 'preview_links' ), $html );
-	}
-
-	/**
-	 * Preview links.
-	 *
-	 * @param array $matches - the matches.
-	 *
-	 * @return string
-	 */
-	public static function preview_links( $matches ) {
-		if ( 0 !== strpos( $matches[2], get_option( 'home' ) ) ) {
-			return $matches[0];
+	public static function skip_stylesheet() {
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$skip_stylesheet = apply_filters( 'safecss_skip_stylesheet', null );
+		if ( $skip_stylesheet !== null ) {
+			return $skip_stylesheet;
 		}
 
-		$link = wp_specialchars_decode( $matches[2] );
-		$link = add_query_arg( 'csspreview', 'true', $link );
-		$link = esc_url( $link );
-		return "href={$matches[1]}$link{$matches[1]}";
-	}
-
-	/**
-	 * Places a black bar above every preview page
-	 */
-	public static function preview_flag() {
-		if ( is_admin() ) {
-			return;
-		}
-
-		$message = esc_html__( 'Preview: changes must be saved or they will be lost', 'jetpack' );
-		/**
-		 * Filter the Preview message displayed on the site when previewing custom CSS, before to save it.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param string $message Custom CSS preview message.
-		 */
-		$message = apply_filters( 'safecss_preview_message', $message );
-
-		$preview_flag_js = "var flag = document.createElement('div');
-		flag.innerHTML = " . wp_json_encode( $message ) . ";
-		flag.style.background = '#FF6600';
-		flag.style.color = 'white';
-		flag.style.textAlign = 'center';
-		flag.style.fontSize = '15px';
-		flag.style.padding = '2px';
-		flag.style.fontFamily = 'sans-serif';
-		document.body.style.paddingTop = '0px';
-		document.body.insertBefore(flag, document.body.childNodes[0]);
-		";
-
-		/**
-		 * Filter the Custom CSS preview message JS styling.
-		 *
-		 * @module custom-css
-		 *
-		 * @since 1.7.0
-		 *
-		 * @param string $preview_flag_js Custom CSS preview message JS styling.
-		 */
-		$preview_flag_js = apply_filters( 'safecss_preview_flag_js', $preview_flag_js );
-		if ( $preview_flag_js ) {
-			$preview_flag_js = '<script type="text/javascript">
-	// <![CDATA[
-	' . $preview_flag_js . '
-	// ]]>
-	</script>';
-		}
-
-		return $preview_flag_js;
-	}
-
-	/**
-	 * Add the additional CSS menu.
-	 */
-	public static function menu() {
-		$title = __( 'Additional CSS', 'jetpack' );
-		$hook  = add_theme_page( $title, $title, 'edit_theme_options', 'editcss', array( 'Jetpack_Custom_CSS', 'admin' ) );
-
-		add_action( 'load-revision.php', array( 'Jetpack_Custom_CSS', 'prettify_post_revisions' ) );
-		add_action( "load-$hook", array( 'Jetpack_Custom_CSS', 'update_title' ) );
-	}
-
-	/**
-	 * Adds a menu item in the appearance section for this plugin's administration
-	 * page. Also adds hooks to enqueue the CSS and JS for the admin page.
-	 */
-	public static function update_title() {
-		global $title;
-		$title = __( 'CSS', 'jetpack' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-	}
-
-	/**
-	 * Prettify the post revision.
-	 */
-	public static function prettify_post_revisions() {
-		add_filter( 'the_title', array( 'Jetpack_Custom_CSS', 'post_title' ), 10, 2 );
-	}
-
-	/**
-	 * Get the post title.
-	 *
-	 * @param string $title - the post title.
-	 * @param int    $post_id - the post ID.
-	 *
-	 * @return string
-	 */
-	public static function post_title( $title, $post_id ) {
-		$post_id = (int) $post_id;
-		if ( ! $post_id ) {
-			return $title;
-		}
-
-		$post = get_post( $post_id );
-		if ( ! $post ) {
-			return $title;
-		}
-
-		if ( 'safecss' !== $post->post_type ) {
-			return $title;
-		}
-
-		return __( 'Custom CSS Stylesheet', 'jetpack' );
-	}
-
-	/**
-	 * Enqueue scripts.
-	 *
-	 * @param string $hook - the hook.
-	 */
-	public static function enqueue_scripts( $hook ) {
-		if ( 'appearance_page_editcss' !== $hook ) {
-			return;
-		}
-
-		wp_enqueue_script( 'postbox' );
-		wp_enqueue_script(
-			'custom-css-editor',
-			Assets::get_file_url_for_environment(
-				'_inc/build/custom-css/custom-css/js/css-editor.min.js',
-				'modules/custom-css/custom-css/js/css-editor.js'
-			),
-			'jquery',
-			'20130325',
-			true
-		);
-		wp_enqueue_style( 'custom-css-editor', plugins_url( 'custom-css/css/css-editor.css', __FILE__ ) ); // phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
-
-		if ( defined( 'SAFECSS_USE_ACE' ) && SAFECSS_USE_ACE ) {
-			wp_register_style( 'jetpack-css-codemirror', plugins_url( 'custom-css/css/codemirror.css', __FILE__ ), array(), '20120905' );
-			wp_enqueue_style( 'jetpack-css-use-codemirror', plugins_url( 'custom-css/css/use-codemirror.css', __FILE__ ), array( 'jetpack-css-codemirror' ), '20120905' );
-
-			wp_register_script( 'jetpack-css-codemirror', plugins_url( 'custom-css/js/codemirror.min.js', __FILE__ ), array(), '3.16', true );
-			wp_enqueue_script(
-				'jetpack-css-use-codemirror',
-				Assets::get_file_url_for_environment(
-					'_inc/build/custom-css/custom-css/js/use-codemirror.min.js',
-					'modules/custom-css/custom-css/js/use-codemirror.js'
-				),
-				array( 'jquery', 'underscore', 'jetpack-css-codemirror' ),
-				'20131009',
-				true
-			);
-		}
-	}
-
-	/**
-	 * Render the saved message.
-	 */
-	public static function saved_message() {
-		echo '<div id="message" class="updated fade"><p><strong>' . esc_html__( 'Stylesheet saved.', 'jetpack' ) . '</strong></p></div>';
-	}
-
-	/**
-	 * Render the admin page.
-	 */
-	public static function admin() {
-		add_meta_box( 'submitdiv', __( 'Publish', 'jetpack' ), array( __CLASS__, 'publish_box' ), 'editcss', 'side' );
-		add_action( 'custom_css_submitbox_misc_actions', array( __CLASS__, 'content_width_settings' ) );
-
-		$safecss_post = self::get_post();
-
-		if ( ! empty( $safecss_post ) && 0 < $safecss_post['ID'] && wp_get_post_revisions( $safecss_post['ID'], array( 'posts_per_page' => 1 ) ) ) {
-			add_meta_box( 'revisionsdiv', __( 'CSS Revisions', 'jetpack' ), array( __CLASS__, 'revisions_meta_box' ), 'editcss', 'side' );
-		}
-		?>
-		<div class="wrap">
-			<?php
-
-			/**
-			 * Fires right before the custom css page begins.
-			 *
-			 * @module custom-css
-			 *
-			 * @since 1.7.0
-			 */
-			do_action( 'custom_design_header' );
-
-			?>
-			<h1><?php esc_html_e( 'CSS Stylesheet Editor', 'jetpack' ); ?></h1>
-			<form id="safecssform" action="" method="post">
-				<?php wp_nonce_field( 'safecss' ); ?>
-				<?php wp_nonce_field( 'meta-box-order', 'meta-box-order-nonce', false ); ?>
-				<?php wp_nonce_field( 'closedpostboxes', 'closedpostboxesnonce', false ); ?>
-				<input type="hidden" name="action" value="save" />
-				<div id="poststuff">
-					<p class="css-support">
-					<?php
-						/**
-						 * Filter the intro text appearing above the Custom CSS Editor.
-						 *
-						 * @module custom-css
-						 *
-						 * @since 1.7.0
-						 *
-						 * @param string $str Intro text appearing above the Custom CSS editor.
-						 */
-						echo apply_filters( // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-							'safecss_intro_text',
-							__(
-								'New to CSS? Start with a <a href="https://www.htmldog.com/guides/css/beginner/" rel="noopener noreferrer" target="_blank">beginner tutorial</a>. Questions?
-		Ask in the <a href="https://wordpress.org/support/forum/themes-and-templates" rel="noopener noreferrer" target="_blank">Themes and Templates forum</a>.',
-								'jetpack'
-							)
-						);
-					?>
-					</p>
-					<p class="css-support"><?php echo esc_html__( 'Note: Custom CSS will be reset when changing themes.', 'jetpack' ); ?></p>
-
-					<div id="post-body" class="metabox-holder columns-2">
-						<div id="post-body-content">
-							<div class="postarea">
-								<textarea id="safecss" name="safecss"
-								<?php
-								if ( SAFECSS_USE_ACE ) {
-									echo ' class="hide-if-js"';
-								}
-								?>
-								><?php echo esc_textarea( self::get_css() ); ?></textarea>
-								<div class="clear"></div>
-							</div>
-						</div>
-						<div id="postbox-container-1" class="postbox-container">
-						<?php do_meta_boxes( 'editcss', 'side', $safecss_post ); ?>
-					</div>
-					</div>
-					<br class="clear" />
-				</div>
-			</form>
-		</div>
-		<?php
-	}
-
-	/**
-	 * Content width setting callback
-	 */
-	public static function content_width_settings() {
-		$safecss_post = self::get_current_revision();
-
-		$custom_content_width = get_post_meta( $safecss_post['ID'], 'content_width', true );
-
-		// If custom content width hasn't been overridden and the theme has a content_width value, use that as a default.
-		if ( $custom_content_width <= 0 && ! empty( $GLOBALS['content_width'] ) ) {
-			$custom_content_width = $GLOBALS['content_width'];
-		}
-
-		if ( ! $custom_content_width || ( isset( $GLOBALS['content_width'] ) && $custom_content_width == $GLOBALS['content_width'] ) ) { // phpcs:ignore Universal.Operators.StrictComparisons.LooseEqual
-			$custom_content_width = '';
-		}
-
-		?>
-		<div class="misc-pub-section">
-			<label><?php esc_html_e( 'Media Width:', 'jetpack' ); ?></label>
-			<span id="content-width-display" data-default-text="<?php esc_attr_e( 'Default', 'jetpack' ); ?>" data-custom-text="
-				<?php
-					// translators: the custom content width.
-					esc_attr_e( '%s px', 'jetpack' );
-				?>
-				">
-				<?php
-					// translators: the custom content width.
-					echo esc_html( $custom_content_width ? sprintf( __( '%s px', 'jetpack' ), $custom_content_width ) : __( 'Default', 'jetpack' ) );
-				?>
-			</span>
-			<a class="edit-content-width hide-if-no-js" href="#content-width"><?php echo esc_html_e( 'Edit', 'jetpack' ); ?></a>
-			<div id="content-width-select" class="hide-if-js">
-				<input type="hidden" name="custom_content_width" id="custom_content_width" value="<?php echo esc_attr( $custom_content_width ); ?>" />
-				<p>
-					<?php
-
-					printf( /* translators: %1$s is replaced with an input field for numbers. */
-						wp_kses_post( __( 'Limit width to %1$s pixels for full size images. (<a href="%2$s" rel="noopener noreferrer" target="_blank">More info</a>.)', 'jetpack' ) ),
-						'<input type="text" id="custom_content_width_visible" value="' . esc_attr( $custom_content_width ) . '" size="4" />',
-						/**
-						 * Filter the Custom CSS limited width's support doc URL.
-						 *
-						 * @module custom-css
-						 *
-						 * @since 2.2.3
-						 *
-						 * @param string $url Custom CSS limited width's support doc URL.
-						 */
-						esc_url(
-							apply_filters( 'safecss_limit_width_link', Redirect::get_url( 'jetpack-support-custom-css', array( 'anchor' => 'limited-width' ) ) )
-						)
-					);
-
-					?>
-				</p>
-				<?php
-
-				if (
-					! empty( $GLOBALS['content_width'] )
-					&& $custom_content_width != $GLOBALS['content_width'] // phpcs:ignore Universal.Operators.StrictComparisons.LooseNotEqual
-				) {
-					$current_theme = wp_get_theme()->Name;
-
-					?>
-					<p>
-					<?php
-					echo esc_html(
-						sprintf(
-							/* translators: %1$s is the theme name, %2$d is an amount of pixels. */
-							_n(
-								'The default content width for the %1$s theme is %2$d pixel.',
-								'The default content width for the %1$s theme is %2$d pixels.',
-								(int) $GLOBALS['content_width'],
-								'jetpack'
-							),
-							$current_theme,
-							(int) $GLOBALS['content_width']
-						)
-					);
-					?>
-					</p>
-					<?php
-				}
-
-				?>
-				<a class="save-content-width hide-if-no-js button" href="#content-width"><?php esc_html_e( 'OK', 'jetpack' ); ?></a>
-				<a class="cancel-content-width hide-if-no-js" href="#content-width"><?php esc_html_e( 'Cancel', 'jetpack' ); ?></a>
-			</div>
-			<script type="text/javascript">
-				jQuery( function ( $ ) {
-					var defaultContentWidth = <?php echo isset( $GLOBALS['content_width'] ) ? wp_json_encode( (int) $GLOBALS['content_width'] ) : 0; ?>;
-
-					$( '.edit-content-width' ).bind( 'click', function ( e ) {
-						e.preventDefault();
-
-						$( '#content-width-select' ).slideDown();
-						$( this ).hide();
-					} );
-
-					$( '.cancel-content-width' ).bind( 'click', function ( e ) {
-						e.preventDefault();
-
-						$( '#content-width-select' ).slideUp( function () {
-							$( '.edit-content-width' ).show();
-							$( '#custom_content_width_visible' ).val( $( '#custom_content_width' ).val() );
-						} );
-					} );
-
-					$( '.save-content-width' ).bind( 'click', function ( e ) {
-						e.preventDefault();
-
-						$( '#content-width-select' ).slideUp();
-
-						var newContentWidth = parseInt( $( '#custom_content_width_visible' ).val(), 10 );
-
-						if ( newContentWidth && newContentWidth != defaultContentWidth ) {
-							$( '#content-width-display' ).text(
-								$( '#content-width-display' )
-									.data( 'custom-text' )
-										.replace( '%s', $( '#custom_content_width_visible' ).val() )
-							);
-						}
-						else {
-							$( '#content-width-display' ).text( $( '#content-width-display' ).data( 'default-text' ) );
-						}
-
-						$( '#custom_content_width' ).val( $( '#custom_content_width_visible' ).val() );
-						$( '.edit-content-width' ).show();
-					} );
-				} );
-			</script>
-		</div>
-		<?php
-	}
-
-	/**
-	 * Render the publish box.
-	 */
-	public static function publish_box() {
-		?>
-		<div id="minor-publishing">
-			<div id="misc-publishing-actions">
-				<?php
-
-				/**
-				 * Filter the array of available Custom CSS preprocessors.
-				 *
-				 * @module custom-css
-				 *
-				 * @since 2.0.3
-				 *
-				 * @param array array() Empty by default.
-				 */
-				$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
-
-				if ( ! empty( $preprocessors ) ) {
-					$safecss_post              = self::get_current_revision();
-					$selected_preprocessor_key = get_post_meta( $safecss_post['ID'], 'custom_css_preprocessor', true );
-					$selected_preprocessor     = isset( $preprocessors[ $selected_preprocessor_key ] ) ? $preprocessors[ $selected_preprocessor_key ] : null;
-
-					?>
-					<div class="misc-pub-section">
-						<label><?php esc_html_e( 'Preprocessor:', 'jetpack' ); ?></label>
-						<span id="preprocessor-display"><?php echo esc_html( $selected_preprocessor ? $selected_preprocessor['name'] : __( 'None', 'jetpack' ) ); ?></span>
-						<a class="edit-preprocessor hide-if-no-js" href="#preprocessor"><?php echo esc_html_e( 'Edit', 'jetpack' ); ?></a>
-						<div id="preprocessor-select" class="hide-if-js">
-							<input type="hidden" name="custom_css_preprocessor" id="custom_css_preprocessor" value="<?php echo esc_attr( $selected_preprocessor_key ); ?>" />
-							<select id="preprocessor_choices">
-								<option value=""><?php esc_html_e( 'None', 'jetpack' ); ?></option>
-								<?php
-
-								foreach ( $preprocessors as $preprocessor_key => $preprocessor ) {
-									?>
-									<option value="<?php echo esc_attr( $preprocessor_key ); ?>" <?php selected( $selected_preprocessor_key, $preprocessor_key ); ?>><?php echo esc_html( $preprocessor['name'] ); ?></option>
-									<?php
-								}
-
-								?>
-							</select>
-							<a class="save-preprocessor hide-if-no-js button" href="#preprocessor"><?php esc_html_e( 'OK', 'jetpack' ); ?></a>
-							<a class="cancel-preprocessor hide-if-no-js" href="#preprocessor"><?php esc_html_e( 'Cancel', 'jetpack' ); ?></a>
-						</div>
-					</div>
-					<?php
-				}
-
-				$safecss_post = self::get_current_revision();
-
-				$add_css = ( get_post_meta( $safecss_post['ID'], 'custom_css_add', true ) !== 'no' );
-
-				?>
-				<div class="misc-pub-section">
-					<label><?php esc_html_e( 'Mode:', 'jetpack' ); ?></label>
-					<span id="css-mode-display"><?php echo esc_html( $add_css ? __( 'Add-on', 'jetpack' ) : __( 'Replacement', 'jetpack' ) ); ?></span>
-					<a class="edit-css-mode hide-if-no-js" href="#css-mode"><?php echo esc_html_e( 'Edit', 'jetpack' ); ?></a>
-					<div id="css-mode-select" class="hide-if-js">
-						<input type="hidden" name="add_to_existing" id="add_to_existing" value="<?php echo $add_css ? 'true' : 'false'; ?>" />
-						<p>
-							<label>
-								<input type="radio" name="add_to_existing_display" value="true" <?php checked( $add_css ); ?>/>
-								<?php echo wp_kses( __( 'Add-on CSS <b>(Recommended)</b>', 'jetpack' ), array( 'b' => array() ) ); ?>
-							</label>
-							<br />
-							<label>
-								<input type="radio" name="add_to_existing_display" value="false" <?php checked( ! $add_css ); ?>/>
-								<?php
-								printf(
-									// translators: the theme's stylesheet URL.
-									wp_kses_post( __( 'Replace <a href="%s">theme\'s CSS</a> <b>(Advanced)</b>', 'jetpack' ) ),
-									/**
-									 * Filter the theme's stylesheet URL.
-									 *
-									 * @module custom-css
-									 *
-									 * @since 1.7.0
-									 *
-									 * @param string $url Active theme's stylesheet URL. Default to get_stylesheet_uri().
-									 */
-									esc_url( apply_filters( 'safecss_theme_stylesheet_url', get_stylesheet_uri() ) )
-								);
-								?>
-							</label>
-						</p>
-						<a class="save-css-mode hide-if-no-js button" href="#css-mode"><?php esc_html_e( 'OK', 'jetpack' ); ?></a>
-						<a class="cancel-css-mode hide-if-no-js" href="#css-mode"><?php esc_html_e( 'Cancel', 'jetpack' ); ?></a>
-					</div>
-				</div>
-				<?php
-
-				/**
-				 * Allows addition of elements to the submit box for custom css on the wp-admin side.
-				 *
-				 * @module custom-css
-				 *
-				 * @since 2.0.3
-				 */
-				do_action( 'custom_css_submitbox_misc_actions' );
-
-				?>
-			</div>
-		</div>
-		<div id="major-publishing-actions">
-			<input type="button" class="button" id="preview" name="preview" value="<?php esc_attr_e( 'Preview', 'jetpack' ); ?>" />
-			<div id="publishing-action">
-				<input type="submit" class="button-primary" id="save" name="save" value="<?php ( self::is_freetrial() ) ? esc_attr_e( 'Save &amp; Buy Upgrade', 'jetpack' ) : esc_attr_e( 'Save Stylesheet', 'jetpack' ); ?>" />
-			</div>
-		</div>
-		<?php
-	}
-
-	/**
-	 * Render metabox listing CSS revisions and the themes that correspond to the revisions.
-	 * Called by safecss_admin
-	 *
-	 * @global $post
-	 * @param array $safecss_post - the safecss array.
-	 * @uses wp_revisions_to_keep
-	 * @uses WP_Query
-	 * @uses wp_post_revision_title
-	 * @uses esc_html
-	 * @uses add_query_arg
-	 * @uses menu_page_url
-	 * @uses wp_reset_query
-	 *
-	 * @todo can this be removed? The revision page seems to work via the customizer now.
-	 */
-	public static function revisions_meta_box( $safecss_post ) {
-
-		$show_all_revisions = isset( $_GET['show_all_rev'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- nothing on the site is changing
-
-		if ( function_exists( 'wp_revisions_to_keep' ) ) {
-			$max_revisions = wp_revisions_to_keep( (object) $safecss_post );
-		} else {
-			$max_revisions = defined( 'WP_POST_REVISIONS' ) && is_numeric( WP_POST_REVISIONS ) ? (int) WP_POST_REVISIONS : 25;
-		}
-
-		$posts_per_page = $show_all_revisions ? $max_revisions : 6;
-
-		$revisions = new WP_Query(
-			array(
-				'posts_per_page' => $posts_per_page,
-				'post_type'      => 'revision',
-				'post_status'    => 'inherit',
-				'post_parent'    => $safecss_post['ID'],
-				'orderby'        => 'date',
-				'order'          => 'DESC',
-			)
-		);
-
-		if ( $revisions->have_posts() ) {
-			?>
-			<ul class="post-revisions">
-			<?php
-
-			global $post;
-
-			while ( $revisions->have_posts() ) :
-				$revisions->the_post();
-
-				?>
-				<li>
-					<?php
-					echo wp_post_revision_title( $post ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-
-					if ( ! empty( $post->post_excerpt ) ) {
-						echo ' (' . esc_html( $post->post_excerpt ) . ')';
-					}
-					?>
-				</li>
-				<?php
-
-			endwhile;
-
-			?>
-			</ul>
-			<?php
-
-			if ( $revisions->found_posts > 6 && ! $show_all_revisions ) {
-				?>
-				<br>
-				<a href="<?php echo esc_url( add_query_arg( 'show_all_rev', 'true', menu_page_url( 'editcss', false ) ) ); ?>"><?php esc_html_e( 'Show all', 'jetpack' ); ?></a>
-				<?php
-			}
-		}
-
-		wp_reset_postdata();
-	}
-
-	/**
-	 * Hook in init at priority 11 to disable custom CSS.
-	 */
-	public static function disable() {
-		remove_action( 'wp_head', array( 'Jetpack_Custom_CSS', 'link_tag' ), 101 );
-		remove_filter( 'stylesheet_uri', array( 'Jetpack_Custom_CSS', 'style_filter' ) );
-	}
-
-	/**
-	 * Reset all aspects of Custom CSS on a theme switch so that changing
-	 * themes is a sure-fire way to get a clean start.
-	 */
-	public static function reset() {
-		$safecss_post_id  = self::save_revision( '' );
-		$safecss_revision = self::get_current_revision();
-
-		update_option( 'safecss_rev', (int) get_option( 'safecss_rev' ) + 1 );
-
-		update_post_meta( $safecss_post_id, 'custom_css_add', 'yes' );
-		update_post_meta( $safecss_post_id, 'content_width', false );
-		update_post_meta( $safecss_post_id, 'custom_css_preprocessor', '' );
-
-		delete_option( 'safecss_add' );
-		delete_option( 'safecss_content_width' );
-
-		update_metadata( 'post', $safecss_revision['ID'], 'custom_css_add', 'yes' );
-		update_metadata( 'post', $safecss_revision['ID'], 'content_width', false );
-		update_metadata( 'post', $safecss_revision['ID'], 'custom_css_preprocessor', '' );
-
-		delete_option( 'safecss_preview_add' );
-	}
-
-	/**
-	 * Checks of we're in the customizer.
-	 *
-	 * @return bool
-	 */
-	public static function is_customizer_preview() {
-		if ( isset( $GLOBALS['wp_customize'] ) ) {
-			return ! $GLOBALS['wp_customize']->is_theme_active();
+		$jetpack_custom_css = get_theme_mod( 'jetpack_custom_css', array() );
+		if ( isset( $jetpack_custom_css['replace'] ) ) {
+			return $jetpack_custom_css['replace'];
 		}
 
 		return false;
 	}
 
 	/**
-	 * Handle minifying CSS.
+	 * Override $content_width in customizer previews.
 	 *
-	 * @param string $css - the CSS.
-	 * @param string $preprocessor - the preprocessor we want to use.
+	 * Runs on `safecss_skip_stylesheet` filter.
+	 *
+	 * @param bool $skip_value Should the stylesheet be skipped.
+	 *
+	 * @return null|bool
+	 */
+	public static function preview_skip_stylesheet( $skip_value ) {
+		global $wp_customize;
+		if ( ! is_customize_preview() ) {
+			return $skip_value;
+		}
+
+		$setting = $wp_customize->get_setting( 'jetpack_custom_css[replace]' );
+		if ( ! $setting ) {
+			return $skip_value;
+		}
+
+		$customized_replace = $setting->post_value();
+		if ( null !== $customized_replace ) {
+			return $customized_replace;
+		}
+
+		return $skip_value;
+	}
+
+	/**
+	 * Add Custom CSS section and controls.
+	 *
+	 * @param WP_Customize_Manager $wp_customize WP_Customize_Manager instance.
+	 */
+	public static function customize_register( $wp_customize ) {
+
+		/**
+		 * SETTINGS.
+		 */
+
+		$wp_customize->add_setting(
+			'jetpack_custom_css[preprocessor]',
+			array(
+				'default'           => '',
+				'transport'         => 'postMessage',
+				'sanitize_callback' => array( __CLASS__, 'sanitize_preprocessor' ),
+			)
+		);
+
+		$wp_customize->add_setting(
+			'jetpack_custom_css[replace]',
+			array(
+				'default'   => false,
+				'transport' => 'refresh',
+			)
+		);
+
+		$wp_customize->add_setting(
+			'jetpack_custom_css[content_width]',
+			array(
+				'default'           => '',
+				'transport'         => 'refresh',
+				'sanitize_callback' => array( __CLASS__, 'intval_base10' ),
+			)
+		);
+
+		// Add custom sanitization to the core css customizer setting.
+		foreach ( $wp_customize->settings() as $setting ) {
+			if ( $setting instanceof WP_Customize_Custom_CSS_Setting ) {
+				add_filter( "customize_sanitize_{$setting->id}", array( __CLASS__, 'sanitize_css_callback' ), 10, 2 );
+			}
+		}
+
+		/**
+		 * CONTROLS.
+		 */
+
+		// Overwrite or Tweak the Core Control.
+		$core_custom_css = $wp_customize->get_control( 'custom_css' );
+		if ( $core_custom_css ) {
+			if ( $core_custom_css instanceof WP_Customize_Code_Editor_Control ) {
+				// In WP 4.9, we let the Core CodeMirror control keep running the show, but hook into it to tweak stuff.
+				$types        = array(
+					'default' => 'text/css',
+					'less'    => 'text/x-less',
+					'sass'    => 'text/x-scss',
+				);
+				$preprocessor = $wp_customize->get_setting( 'jetpack_custom_css[preprocessor]' )->value();
+				if ( isset( $types[ $preprocessor ] ) ) {
+					$core_custom_css->code_type = $types[ $preprocessor ];
+				}
+			} else {
+				// Core < 4.9 Fallback
+				$core_custom_css->type = 'jetpackCss';
+			}
+		}
+
+		$wp_customize->selective_refresh->add_partial(
+			'custom_css',
+			array(
+				'type'                => 'custom_css',
+				'selector'            => '#wp-custom-css',
+				'container_inclusive' => false,
+				'fallback_refresh'    => false,
+				'settings'            => array(
+					'custom_css[' . $wp_customize->get_stylesheet() . ']',
+					'jetpack_custom_css[preprocessor]',
+				),
+				'render_callback'     => array( __CLASS__, 'echo_custom_css_partial' ),
+			)
+		);
+
+		$wp_customize->add_control(
+			'wpcom_custom_css_content_width_control',
+			array(
+				'type'     => 'text',
+				'label'    => __( 'Media Width', 'jetpack' ),
+				'section'  => 'custom_css',
+				'settings' => 'jetpack_custom_css[content_width]',
+			)
+		);
+
+		$wp_customize->add_control(
+			'jetpack_css_mode_control',
+			array(
+				'type'     => 'checkbox',
+				'label'    => __( 'Don\'t use the theme\'s original CSS.', 'jetpack' ),
+				'section'  => 'custom_css',
+				'settings' => 'jetpack_custom_css[replace]',
+			)
+		);
+
+		/**
+		 * An action to grab on to if another Jetpack Module would like to add its own controls.
+		 *
+		 * @module custom-css
+		 *
+		 * @since 4.4.2
+		 *
+		 * @param $wp_customize The WP_Customize object.
+		 */
+		do_action( 'jetpack_custom_css_customizer_controls', $wp_customize );
+
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+		if ( ! empty( $preprocessors ) ) {
+			$preprocessor_choices = array(
+				'' => __( 'None', 'jetpack' ),
+			);
+
+			foreach ( $preprocessors as $preprocessor_key => $processor ) {
+				$preprocessor_choices[ $preprocessor_key ] = $processor['name'];
+			}
+
+			$wp_customize->add_control(
+				'jetpack_css_preprocessors_control',
+				array(
+					'type'     => 'select',
+					'choices'  => $preprocessor_choices,
+					'label'    => __( 'Preprocessor', 'jetpack' ),
+					'section'  => 'custom_css',
+					'settings' => 'jetpack_custom_css[preprocessor]',
+				)
+			);
+		}
+
+	}
+
+	/**
+	 * The callback to handle sanitizing the CSS. Takes different arguments, hence the proxy function.
+	 *
+	 * @param mixed $css Value of the setting.
+	 *
+	 * @return mixed|string
+	 */
+	public static function sanitize_css_callback( $css ) {
+		global $wp_customize;
+		return self::sanitize_css(
+			$css,
+			array(
+				'preprocessor' => $wp_customize->get_setting( 'jetpack_custom_css[preprocessor]' )->value(),
+			)
+		);
+	}
+
+	/**
+	 * Flesh out for wpcom.
+	 *
+	 * @todo
+	 *
+	 * @return bool
+	 */
+	public static function is_freetrial() {
+		return false;
+	}
+
+	/**
+	 * Flesh out for wpcom.
+	 *
+	 * @todo
+	 *
+	 * @return bool
+	 */
+	public static function is_preview() {
+		return false;
+	}
+
+	/**
+	 * Output the custom css for customize preview.
+	 *
+	 * @param string $css Custom CSS content.
+	 *
+	 * @return mixed
+	 */
+	public static function customize_preview_wp_get_custom_css( $css ) {
+		global $wp_customize;
+
+		$preprocessor = $wp_customize->get_setting( 'jetpack_custom_css[preprocessor]' )->value();
+
+		// If it's empty, just return.
+		if ( empty( $preprocessor ) ) {
+			return $css;
+		}
+
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+		if ( isset( $preprocessors[ $preprocessor ] ) ) {
+			return call_user_func( $preprocessors[ $preprocessor ]['callback'], $css );
+		}
+
+		return $css;
+	}
+
+	/**
+	 * Add CSS preprocessing to our CSS if it is supported.
+	 *
+	 * @param mixed                $css     Value of the setting.
+	 * @param WP_Customize_Setting $setting WP_Customize_Setting instance.
 	 *
 	 * @return string
 	 */
-	public static function minify( $css, $preprocessor = '' ) {
-		if ( ! $css ) {
-			return '';
+	public static function customize_value_custom_css( $css, $setting ) {
+		// Find the current preprocessor.
+		$jetpack_custom_css = get_theme_mod( 'jetpack_custom_css', array() );
+		if ( isset( $jetpack_custom_css['preprocessor'] ) ) {
+			$preprocessor = $jetpack_custom_css['preprocessor'];
 		}
 
-		if ( $preprocessor ) {
-			/** This filter is documented in modules/custom-css/custom-css.php */
-			$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
-
-			if ( isset( $preprocessors[ $preprocessor ] ) ) {
-				$css = call_user_func( $preprocessors[ $preprocessor ]['callback'], $css );
-			}
+		// If it's not supported, just return.
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+		if ( ! isset( $preprocessors[ $preprocessor ] ) ) {
+			return $css;
 		}
 
-		safecss_class();
-		$csstidy           = new csstidy();
-		$csstidy->optimise = new safecss( $csstidy );
+		// Swap it for the `post_content_filtered` instead.
+		$post = wp_get_custom_css_post( $setting->stylesheet );
+		if ( $post && ! empty( $post->post_content_filtered ) ) {
+			$css = $post->post_content_filtered;
+		}
 
-		$csstidy->set_cfg( 'remove_bslash', false );
-		$csstidy->set_cfg( 'compress_colors', true );
-		$csstidy->set_cfg( 'compress_font-weight', true );
-		$csstidy->set_cfg( 'remove_last_;', true );
-		$csstidy->set_cfg( 'case_properties', true );
-		$csstidy->set_cfg( 'discard_invalid_properties', true );
-		$csstidy->set_cfg( 'css_level', 'CSS3.0' );
-		$csstidy->set_cfg( 'template', 'highest' );
-		$csstidy->parse( $css );
-
-		return $csstidy->print->plain();
+		return $css;
 	}
 
 	/**
-	 * When restoring a SafeCSS post revision, also copy over the
-	 * content_width and custom_css_add post metadata.
+	 * Store the original pre-processed CSS in `post_content_filtered`
+	 * and then store processed CSS in `post_content`.
 	 *
-	 * @param int $_post_id - the post ID.
-	 * @param int $_revision_id - the revision ID.
-	 */
-	public static function restore_revision( $_post_id, $_revision_id ) {
-		$_post = get_post( $_post_id );
-
-		if ( 'safecss' !== $_post->post_type ) {
-			return;
-		}
-
-		$safecss_revision = self::get_current_revision();
-
-		$content_width  = get_post_meta( $_revision_id, 'content_width', true );
-		$custom_css_add = get_post_meta( $_revision_id, 'custom_css_add', true );
-		$preprocessor   = get_post_meta( $_revision_id, 'custom_css_preprocessor', true );
-
-		update_metadata( 'post', $safecss_revision['ID'], 'content_width', $content_width );
-		update_metadata( 'post', $safecss_revision['ID'], 'custom_css_add', $custom_css_add );
-		update_metadata( 'post', $safecss_revision['ID'], 'custom_css_preprocessor', $preprocessor );
-
-		delete_option( 'safecss_add' );
-		delete_option( 'safecss_content_width' );
-
-		update_post_meta( $_post->ID, 'content_width', $content_width );
-		update_post_meta( $_post->ID, 'custom_css_add', $custom_css_add );
-		update_post_meta( $_post->ID, 'custom_css_preprocessor', $preprocessor );
-
-		delete_option( 'safecss_preview_add' );
-	}
-
-	/**
-	 * Migration routine for moving safecss from wp_options to wp_posts to support revisions
-	 */
-	public static function upgrade() {
-		$css = get_option( 'safecss' );
-
-		if ( get_option( 'safecss_revision_migrated' ) ) {
-			return false;
-		}
-
-		// Check if CSS is stored in wp_options
-		if ( $css ) {
-			// Remove the async actions from publish_post
-			remove_action( 'publish_post', 'queue_publish_post' );
-
-			$post                 = array();
-			$post['post_content'] = $css;
-			$post['post_title']   = 'safecss';
-			$post['post_status']  = 'publish';
-			$post['post_type']    = 'safecss';
-
-			// Insert the CSS into wp_posts
-			$post_id = wp_insert_post( $post );
-			// Check for errors
-			if ( ! $post_id || is_wp_error( $post_id ) ) {
-				die( $post_id->get_error_message() ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-			}
-
-			// Delete safecss option
-			delete_option( 'safecss' );
-		}
-
-		unset( $css );
-
-		// Check if we have already done this
-		if ( ! get_option( 'safecss_revision_migrated' ) ) {
-			define( 'DOING_MIGRATE', true );
-
-			// Get hashes of safecss post and current revision
-			$safecss_post = self::get_post();
-
-			if ( empty( $safecss_post ) ) {
-				return;
-			}
-
-			$safecss_post_hash = md5( $safecss_post['post_content'] );
-			$current_revision  = self::get_current_revision();
-
-			if ( null === $current_revision ) {
-				return;
-			}
-
-			$current_revision_hash = md5( $current_revision['post_content'] );
-
-			// If hashes are not equal, set safecss post with content from current revision
-			if ( $safecss_post_hash !== $current_revision_hash ) {
-				self::save_revision( $current_revision['post_content'] );
-				// Reset post_content to display the migrated revsion
-				$safecss_post['post_content'] = $current_revision['post_content'];
-			}
-
-			// Set option so that we dont keep doing this
-			update_option( 'safecss_revision_migrated', time() );
-		}
-
-		$newest_safecss_post = self::get_current_revision();
-
-		if ( $newest_safecss_post ) {
-			if ( get_option( 'safecss_content_width' ) ) {
-				// Add the meta to the post and the latest revision.
-				update_post_meta( $newest_safecss_post['ID'], 'content_width', get_option( 'safecss_content_width' ) );
-				update_metadata( 'post', $newest_safecss_post['ID'], 'content_width', get_option( 'safecss_content_width' ) );
-
-				delete_option( 'safecss_content_width' );
-			}
-
-			if ( get_option( 'safecss_add' ) ) {
-				update_post_meta( $newest_safecss_post['ID'], 'custom_css_add', get_option( 'safecss_add' ) );
-				update_metadata( 'post', $newest_safecss_post['ID'], 'custom_css_add', get_option( 'safecss_add' ) );
-
-				delete_option( 'safecss_add' );
-			}
-		}
-	}
-
-	/**
-	 * Adds a filter to the redirect location in `wp-admin/revisions.php`.
-	 */
-	public static function add_revision_redirect() {
-		add_filter( 'wp_redirect', array( __CLASS__, 'revision_redirect' ) );
-	}
-
-	/**
-	 * Filters the redirect location in `wp-admin/revisions.php`.
+	 * @param array  $args Content post args.
+	 * @param string $css  Original CSS being updated.
 	 *
-	 * @param string $location The path to redirect to.
-	 * @return string
+	 * @return mixed
 	 */
-	public static function revision_redirect( $location ) {
-		$post = get_post();
-
-		if ( ! empty( $post->post_type ) && 'safecss' === $post->post_type ) {
-			$location = 'themes.php?page=editcss';
-
-			if ( 'edit.php' === $location ) {
-				$location = '';
-			}
+	public static function customize_update_custom_css_post_content_args( $args, $css ) {
+		// Find the current preprocessor.
+		$jetpack_custom_css = get_theme_mod( 'jetpack_custom_css', array() );
+		if ( empty( $jetpack_custom_css['preprocessor'] ) ) {
+			return $args;
 		}
 
-		return $location;
+		$preprocessor = $jetpack_custom_css['preprocessor'];
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+
+		// If it's empty, just return.
+		if ( empty( $preprocessor ) ) {
+			return $args;
+		}
+
+		if ( isset( $preprocessors[ $preprocessor ] ) ) {
+			$args['post_content_filtered'] = $css;
+			$args['post_content']          = call_user_func( $preprocessors[ $preprocessor ]['callback'], $css );
+		}
+
+		return $args;
 	}
 
 	/**
-	 * The revision post link.
+	 * Filter to handle the processing of preprocessed css on save.
 	 *
-	 * @param string $post_link - the post link.
-	 * @param int    $post_id - the post ID.
-	 * @param string $context - the context.
+	 * @param array $args Custom CSS options.
 	 *
-	 * @return string
+	 * @return mixed
 	 */
-	public static function revision_post_link( $post_link, $post_id, $context ) {
-		$post_id = (int) $post_id;
-		if ( ! $post_id ) {
-			return $post_link;
+	public static function update_custom_css_data( $args ) {
+		// Find the current preprocessor.
+		$jetpack_custom_css = get_theme_mod( 'jetpack_custom_css', array() );
+		if ( empty( $jetpack_custom_css['preprocessor'] ) ) {
+			return $args;
 		}
 
-		$post = get_post( $post_id );
-		if ( ! $post ) {
-			return $post_link;
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+		$preprocessor  = $jetpack_custom_css['preprocessor'];
+
+		// If we have a preprocessor specified ...
+		if ( isset( $preprocessors[ $preprocessor ] ) ) {
+			// And no other preprocessor has run ...
+			if ( empty( $args['preprocessed'] ) ) {
+				$args['preprocessed'] = $args['css'];
+				$args['css']          = call_user_func( $preprocessors[ $preprocessor ]['callback'], $args['css'] );
+			} else {
+				trigger_error( 'Jetpack CSS Preprocessor specified, but something else has already modified the argument.', E_USER_WARNING ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_trigger_error
+			}
 		}
 
-		if ( 'safecss' !== $post->post_type ) {
-			return $post_link;
-		}
-
-		$post_link = admin_url( 'themes.php?page=editcss' );
-
-		if ( 'display' === $context ) {
-			return esc_url( $post_link );
-		}
-
-		return esc_url_raw( $post_link );
+		return $args;
 	}
 
 	/**
 	 * When on the edit screen, make sure the custom content width
 	 * setting is applied to the large image size.
 	 *
-	 * @param array  $dims - the width and height dimensions.
-	 * @param string $size - the size.
-	 * @param string $context - the context in which we're applying dimensions.
+	 * @param array  $dims    Array of image dimensions (width and height).
+	 * @param string $size    Size of the resulting image.
+	 * @param null   $context Context the image is being resized for. `edit` or `display`.
 	 *
 	 * @return array
 	 */
@@ -1806,86 +973,194 @@ class Jetpack_Custom_CSS {
 	/**
 	 * Override the content_width with a custom value if one is set.
 	 *
-	 * @param int $content_width - the content width in pixels.
+	 * @param int $content_width Content Width value to be updated.
 	 *
 	 * @return int
 	 */
 	public static function jetpack_content_width( $content_width ) {
 		$custom_content_width = 0;
 
-		if ( self::is_preview() ) {
-			$safecss_post         = self::get_current_revision();
-			$custom_content_width = (int) get_post_meta( $safecss_post['ID'], 'content_width', true );
-		} elseif ( ! self::is_freetrial() ) {
-			$custom_css_post_id = self::post_id();
-			if ( $custom_css_post_id ) {
-				$custom_content_width = (int) get_post_meta( $custom_css_post_id, 'content_width', true );
-			}
+		$jetpack_custom_css = get_theme_mod( 'jetpack_custom_css', array() );
+		if ( isset( $jetpack_custom_css['content_width'] ) ) {
+			$custom_content_width = $jetpack_custom_css['content_width'];
 		}
 
 		if ( $custom_content_width > 0 ) {
-			$content_width = $custom_content_width;
+			return $custom_content_width;
 		}
 
 		return $content_width;
 	}
-}
 
-/**
- * The Safe CSS Class.
- */
-class Jetpack_Safe_CSS { // phpcs:ignore Generic.Files.OneObjectStructurePerFile.MultipleFound, Generic.Classes.OpeningBraceSameLine.ContentAfterBrace
 	/**
-	 * Filter attriburtes.
-	 *
-	 * @param string $css - the CSS.
-	 * @param string $element - the HTML element.
-	 *
-	 * @return string
+	 * Currently this filter function gets called on
+	 * 'template_redirect' action and
+	 * 'admin_init' action
 	 */
-	public static function filter_attr( $css, $element = 'div' ) {
-		safecss_class();
-
-		$css = $element . ' {' . $css . '}';
-
-		$csstidy           = new csstidy();
-		$csstidy->optimise = new safecss( $csstidy );
-		$csstidy->set_cfg( 'remove_bslash', false );
-		$csstidy->set_cfg( 'compress_colors', false );
-		$csstidy->set_cfg( 'compress_font-weight', false );
-		$csstidy->set_cfg( 'discard_invalid_properties', true );
-		$csstidy->set_cfg( 'merge_selectors', false );
-		$csstidy->set_cfg( 'remove_last_;', false );
-		$csstidy->set_cfg( 'css_level', 'CSS3.0' );
-
-		// Turn off css shorthands and leading zero removal as it breaks block validation.
-		$csstidy->set_cfg( 'optimise_shorthands', 0 );
-		$csstidy->set_cfg( 'preserve_leading_zeros', true );
-
-		$css = preg_replace( '/\\\\([0-9a-fA-F]{4})/', '\\\\\\\\$1', $css );
-		$css = wp_kses_split( $css, array(), array() );
-		$csstidy->parse( $css );
-
-		$css = $csstidy->print->plain();
-
-		$css = str_replace( array( "\n", "\r", "\t" ), '', $css );
-
-		preg_match( "/^{$element}\s*{(.*)}\s*$/", $css, $matches );
-
-		if ( empty( $matches[1] ) ) {
-			return '';
+	public static function set_content_width() {
+		// Don't apply this filter on the Edit CSS page.
+		if ( isset( $_GET['page'] ) && 'editcss' === $_GET['page'] && is_admin() ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- nothing is changed on the site.
+			return;
 		}
 
-		return $matches[1];
+		$GLOBALS['content_width'] = Jetpack::get_content_width();
+	}
+
+	/**
+	 * Make sure the preprocessor we're saving is one we know about.
+	 *
+	 * @param string $preprocessor The preprocessor to sanitize.
+	 *
+	 * @return null|string
+	 */
+	public static function sanitize_preprocessor( $preprocessor ) {
+		/** This filter is documented in modules/custom-css/custom-css.php */
+		$preprocessors = apply_filters( 'jetpack_custom_css_preprocessors', array() );
+		if ( empty( $preprocessor ) || array_key_exists( $preprocessor, $preprocessors ) ) {
+			return $preprocessor;
+		}
+		return null;
+	}
+
+	/**
+	 * Get the base10 intval.
+	 *
+	 * This is used as a setting's sanitize_callback; we can't use just plain
+	 * intval because the second argument is not what intval() expects.
+	 *
+	 * @access public
+	 *
+	 * @param mixed $value Number to convert.
+	 * @return int Integer.
+	 */
+	public static function intval_base10( $value ) {
+		return (int) $value;
+	}
+
+	/**
+	 * Add a footer action on revision.php to print some customizations for the theme switcher.
+	 */
+	public static function load_revision_php() {
+		add_action( 'admin_footer', array( __CLASS__, 'revision_admin_footer' ) );
+	}
+
+	/**
+	 * Print the theme switcher on revision.php and move it into place.
+	 */
+	public static function revision_admin_footer() {
+		$post = get_post();
+		if ( 'custom_css' !== $post->post_type ) {
+			return;
+		}
+		$stylesheet = $post->post_title;
+		?>
+<script type="text/html" id="tmpl-other-themes-switcher">
+		<?php self::revisions_switcher_box( $stylesheet ); ?>
+</script>
+<style>
+.other-themes-wrap {
+	float: right;
+	background-color: #fff;
+	-webkit-box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+	box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+	padding: 5px 10px;
+	margin-bottom: 10px;
+}
+.other-themes-wrap label {
+	display: block;
+	margin-bottom: 10px;
+}
+.other-themes-wrap select {
+	float: left;
+	width: 77%;
+}
+.other-themes-wrap button {
+	float: right;
+	width: 20%;
+}
+.revisions {
+	clear: both;
+}
+/* Hide the back-to-post link */
+.long-header + a {
+	display: none;
+}
+</style>
+<script>
+(function($){
+	var switcher = $('#tmpl-other-themes-switcher').html(),
+		qty = $( switcher ).find('select option').length,
+		$switcher;
+
+	if ( qty >= 3 ) {
+		$('h1.long-header').before( switcher );
+		$switcher = $('.other-themes-wrap');
+		$switcher.find('button').on('click', function(e){
+			e.preventDefault();
+			if ( $switcher.find('select').val() ) {
+				window.location.href = $switcher.find('select').val();
+			}
+		})
+	}
+})(jQuery);
+</script>
+		<?php
+	}
+
+	/**
+	 * The HTML for the theme revision switcher box.
+	 *
+	 * @param string $stylesheet Stylesheet name.
+	 */
+	public static function revisions_switcher_box( $stylesheet = '' ) {
+		$themes = self::get_all_themes_with_custom_css();
+		?>
+		<div class="other-themes-wrap">
+			<label for="other-themes"><?php esc_html_e( 'Select another theme to view its custom CSS.', 'jetpack' ); ?></label>
+			<select id="other-themes">
+				<option value=""><?php esc_html_e( 'Select a theme&hellip;', 'jetpack' ); ?></option>
+				<?php
+				foreach ( $themes as $theme_stylesheet => $data ) {
+					$revisions = wp_get_post_revisions( $data['post']->ID, array( 'posts_per_page' => 1 ) );
+					if ( ! $revisions ) {
+						?>
+						<option value="<?php echo esc_url( add_query_arg( 'id', $data['post']->ID, menu_page_url( 'editcss', 0 ) ) ); ?>" <?php disabled( $stylesheet, $theme_stylesheet ); ?>>
+							<?php echo esc_html( $data['label'] ); ?>
+							<?php
+								// translators: how long ago the stylesheet was modified.
+								printf( esc_html__( '(modified %s ago)', 'jetpack' ), human_time_diff( strtotime( $data['post']->post_modified_gmt ) ) ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+							?>
+						</option>
+						<?php
+						continue;
+					}
+					$revision = array_shift( $revisions );
+					?>
+					<option value="<?php echo esc_url( get_edit_post_link( $revision->ID ) ); ?>" <?php disabled( $stylesheet, $theme_stylesheet ); ?>>
+						<?php echo esc_html( $data['label'] ); ?>
+						<?php
+							// translators: how long ago the stylesheet was modified.
+							printf( esc_html__( '(modified %s ago)', 'jetpack' ), human_time_diff( strtotime( $data['post']->post_modified_gmt ) ) );  // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+						?>
+					</option>
+					<?php
+				}
+				?>
+			</select>
+			<button class="button" id="other_theme_custom_css_switcher"><?php esc_html_e( 'Switch', 'jetpack' ); ?></button>
+		</div>
+		<?php
 	}
 }
 
+Jetpack_Custom_CSS_Enhancements::add_hooks();
+
 if ( ! function_exists( 'safecss_class' ) ) :
 	/**
-	 * Setup safecss class.
+	 * Load in the class only when needed.  Makes lighter load by having one less class in memory.
 	 */
 	function safecss_class() {
-		// Wrapped so we don't need the parent class just to load the plugin
+		// Wrapped so we don't need the parent class just to load the plugin.
 		if ( class_exists( 'safecss' ) ) {
 			return;
 		}
@@ -1893,43 +1168,27 @@ if ( ! function_exists( 'safecss_class' ) ) :
 		require_once __DIR__ . '/csstidy/class.csstidy.php';
 
 		/**
-		 * Safe CSS Class.
+		 * Class safecss
 		 */
 		class safecss extends csstidy_optimise { // phpcs:ignore
 
 			/**
-			 * Add action to fire after parsing CSS.
+			 * Optimises $css after parsing.
 			 */
 			public function postparse() { // phpcs:ignore MediaWiki.Usage.NestedFunctions.NestedFunction
 
-				/**
-				 * Fires after parsing the css.
-				 *
-				 * @module custom-css
-				 *
-				 * @since 1.8.0
-				 *
-				 * @param obj $this CSSTidy object.
-				 */
+				/** This action is documented in modules/custom-css/custom-css.php */
 				do_action( 'csstidy_optimize_postparse', $this );
 
 				return parent::postparse();
 			}
 
 			/**
-			 * Handle subvalue action.
+			 * Optimises a sub-value.
 			 */
 			public function subvalue() { // phpcs:ignore MediaWiki.Usage.NestedFunctions.NestedFunction
 
-				/**
-				 * Fires before optimizing the Custom CSS subvalue.
-				 *
-				 * @module custom-css
-				 *
-				 * @since 1.8.0
-				 *
-				 * @param obj $this CSSTidy object.
-				 */
+				/** This action is documented in modules/custom-css/custom-css.php */
 				do_action( 'csstidy_optimize_subvalue', $this );
 
 				return parent::subvalue();
@@ -1937,18 +1196,3 @@ if ( ! function_exists( 'safecss_class' ) ) :
 		}
 	}
 endif;
-
-if ( ! function_exists( 'safecss_filter_attr' ) ) {
-
-	/**
-	 * Filter safecss attriburtes.
-	 *
-	 * @param string $css - the CSS.
-	 * @param string $element - the HTML element.
-	 */
-	function safecss_filter_attr( $css, $element = 'div' ) {
-		return Jetpack_Safe_CSS::filter_attr( $css, $element );
-	}
-}
-
-require_once __DIR__ . '/custom-css/preprocessors.php';
